% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{aa2codon}
\alias{aa2codon}
\title{Find the DNA to code for a given amino acid}
\usage{
aa2codon(aa, type = "code", regex = TRUE)
}
\arguments{
\item{aa}{amino acid to convert to dna}

\item{type}{code or abbr or name}

\item{regex}{if true return a (X|Y) regex of codons else return vector}
}
\value{
A vector of codons if regex=FALSE else a regular expression matching the possible codons
}
\description{
Find the DNA to code for a given amino acid
}

