% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{aa2codons}
\alias{aa2codons}
\title{Find the DNA to code for a given amino acid}
\usage{
aa2codons(aas, type = "code", regex = TRUE)
}
\arguments{
\item{type}{code or abbr or name}

\item{regex}{if true return a (X|Y) regex of codons else return vector}

\item{aa}{single amino acids to convert to dna}
}
\value{
A list of vectors of codons if regex=FALSE else a regular expression matching the possible codons
}
\description{
Find the DNA to code for a given amino acid
}
\examples{
aa2codons(c('A','T','X'))
aa2codons(c('A','T','X'),regex=FALSE)
aa2codons(c('Glu','Ala','Xxx'),type='abbr')
}

