% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{add_gate_binding_on_neuron_hje}
\alias{add_gate_binding_on_neuron_hje}
\title{Add a gate binding into a neuron}
\usage{
add_gate_binding_on_neuron_hje(neuron, binding)
}
\arguments{
\item{neuron}{The neuron which will have a gate binding added.}

\item{binding}{A list representing the binding. The structure
can be anything.}
}
\value{
The neuron passed as parameter but with the binding in it.
}
\description{
Use this function to add a binding into a list called \code{gate_bindings}
within the neuron. this list will be create if it doesn't exists
}
