% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{addSnpsToGenome}
\alias{addSnpsToGenome}
\title{Add SNPs and indels to a reference sequence}
\usage{
addSnpsToGenome(ref, snps = NULL, insertions = NULL, deletions = NULL)
}
\arguments{
\item{ref}{a string giving a reference sequence to be edited}

\item{snps}{a character vector giving SNPs in the format reference nucleotide, 1-based position and mutant nucleotide with no spaces .e.g. A3T would mean that the reference CCACC changed to CCTCC}

\item{insertions}{a character vector giving insertions in the format 1-based position of the nucleotide prior to the insertion and the inserted nucleotides with no space e.g. 3TT would mean that the reference CCACC changed to CCATTCC}

\item{deletions}{a character vector giving insertions in the format 1-based position of the first nucleotide of the deletions, dash and position of the last nucleotide of the insertion with no space e.g. 2-4 would mean that the reference CCACC changed to C---C}
}
\value{
a single character string containing the reference modified to contain the specified SNPs and indels
}
\description{
Add SNPs and indels to a reference sequence
}
\examples{
addSnpsToGenome('ABCDEFGHIJKL',c('B2Z','D4Y'))
addSnpsToGenome('ABCDEFGHIJKL',deletions=c('1-4','6'))
addSnpsToGenome('ABCDEFGHIJKL',insertions=c('4XYZ','6VW'))
}
\seealso{
\code{\link{degap}}
}
