% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{adjustWindow}
\alias{adjustWindow}
\title{Convenience function for selecting multiple elements from a matrix by x,y position
convenience function to resize R console window}
\usage{
adjustWindow(width = as.integer(Sys.getenv("COLUMNS")))
}
\arguments{
\item{width}{width to adjust window to}
}
\description{
Convenience function for selecting multiple elements from a matrix by x,y position
convenience function to resize R console window
}
