% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{allSameLength}
\alias{allSameLength}
\title{Convenience function to check all args are same length}
\usage{
allSameLength(...)
}
\arguments{
\item{...}{arguments to check the lengths of}
}
\value{
logical indicating whether all input are the same length
}
\description{
Convenience function to check all args are same length
}

