% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{ambiguous2regex}
\alias{ambiguous2regex}
\title{Converts ambiguous dna to an appropriate regular expression}
\usage{
ambiguous2regex(dna)
}
\arguments{
\item{dna}{input sequence with ambiguous bases}
}
\value{
regular expression to match the ambiguous dna
}
\description{
Converts ambiguous dna to an appropriate regular expression
}
\examples{
dnaRegex<-ambiguous2regex('ANRT')
grepl(dnaRegex,c('ATGT','ACAT','AGTT'))
}
\references{
\url{https://en.wikipedia.org/wiki/Nucleic_acid_notation}
}

