% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\docType{data}
\name{ambiguousBaseCodes}
\alias{ambiguousBaseCodes}
\title{Base codes used to indicate ambiguous bases}
\format{An object of class \code{character} of length 11.}
\usage{
ambiguousBaseCodes
}
\description{
Base codes used to indicate ambiguous bases
}
\keyword{datasets}

