% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_behavior}
\alias{analyze_behavior}
\title{This function returns the concentration derivative of each species}
\usage{
analyze_behavior(
  species,
  ci,
  reactions,
  ki,
  time_points = NULL,
  behavior = NULL
)
}
\arguments{
\item{species}{A vector with the species of the reaction. The order of
this vector is important because it will define the
column order of the returned behavior.}

\item{ci}{A vector specifying the initial concentrations of the
\code{species} specified, in order.}

\item{reactions}{A vector with the reactions of the CRN. If a reaction has
has only reactants that are non in \code{species}, this
reaction will be treated as \code{0 -> products}. Furthermore,
this function treats a reaction \code{non_registered_species + registered_species -> products} equally to
\code{registered_species -> products}, ignoring the species non
registered on the \code{species} vector.}

\item{ki}{A vector defining the constant rate of each reaction
in \code{reactions}, in order.}

\item{time_points}{A vector of indexes (representing multiple points in
time) used for access lines of \code{behavior}.}

\item{behavior}{The data returned by \code{\link{react}()}.}
}
\value{
A data frame with the derivatives. To access the derivative
of a species \code{'A'}, you just have to access \code{df['A']}.
}
\description{
This function can be used for study what is impacting each species and
how much. this is useful to analyse medium size (dozens of reactions) CRNs.
all parameters follows the parameters of \code{\link{react}()}, except
the optional \code{time_point} and \code{behavior}. If a \code{time_point} is passed,
a \code{behavior} must be passed as well. If both parameters are set,
this functions returns the concentration of each species
at a specific point in time within the derivative.
}
