% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{anyArgsNA}
\alias{anyArgsNA}
\title{Convenience function to check whether any  args have NAs}
\usage{
anyArgsNA(..., recursive = FALSE)
}
\arguments{
\item{...}{arguments to check if contain NA}

\item{recursive}{A logical indicating whether to recursively look inside arguments for NAs}
}
\value{
logical indicating whether any argument contains NA
}
\description{
Convenience function to check whether any  args have NAs
}

