% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{anyNa}
\alias{anyNa}
\title{Convenience function to check whether any  args have NAs}
\usage{
anyNa(...)
}
\arguments{
\item{...}{arguments to check if contain NA}
}
\value{
logical indicating whether any argument contains NA
}
\description{
Convenience function to check whether any  args have NAs
}

