% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsd.R
\name{arg_list}
\alias{arg_list}
\title{Get the arguments of the current function}
\usage{
arg_list()
}
\value{
A list with the name of the arguments being the name of the
element, and the argument's content being the content of
the element.
}
\description{
Use this function to get the arguments of the current function.
}
