% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{arrow}
\alias{arrow}
\title{Find coords for plotting horizontal arrows}
\usage{
arrow(tail, head, y, arrowLength = diff(graphics::par("usr")[1:2]) *
  0.05, shaft = 0.2, point = 0.4, concat = TRUE)
}
\arguments{
\item{tail}{x coordinate for tail of the arrow}

\item{head}{x coordinate for the head of the arrow}

\item{y}{y position for middle of block}

\item{arrowLength}{arrow length in usr coords}

\item{shaft}{half of shaft thickness in usr coords}

\item{point}{half of arrow thickness in usr coords}

\item{concat}{logical indicating whether to concatenate multiple arrows into one data frame separated by NAs (e.g. ready for poly)}
}
\value{
a data.frame with columns x and y specifying coordinates for arrows
}
\description{
Find coords for plotting horizontal arrows
}
\examples{
plot(1:10)
polygon(arrow(1:5,3:7*1.5,2:6))
polygon(arrow(3:6,1:4/1.5,7:10))
polygon(arrow(7:10,7:10+0:3*.1,7:10))
polygon(arrow(6:9,6:9-0:3*.1,7:10))
}
