% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{bases2ambiguous}
\alias{bases2ambiguous}
\title{Convert set of single bases to ambigous code}
\usage{
bases2ambiguous(bases)
}
\arguments{
\item{bases}{vector of single character base strings}
}
\description{
Convert set of single bases to ambigous code
}

