% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{binary2range}
\alias{binary2range}
\title{Find covered ranges in binary data}
\usage{
binary2range(index)
}
\arguments{
\item{index}{logical vector containing TRUE in the regions of interest}
}
\value{
data.frame with rows for each contiguous region of interest with columns start and end of regions
}
\description{
Find covered ranges in binary data
}
\examples{
binary2range(c(FALSE,TRUE,FALSE,TRUE,TRUE,TRUE))
}
