% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlat.R
\name{blatReadsVsRefs}
\alias{blatReadsVsRefs}
\title{Make a 2bit file from one set of reads and blat another set against it}
\usage{
blatReadsVsRefs(reads, refs, outFile, faToTwoBit = "faToTwoBit",
  tmpDir = tempfile("dir"), ...)
}
\arguments{
\item{reads}{vector of query reads with names}

\item{refs}{vector of reference reads with names}

\item{outFile}{file to write blat to (if ends in .gz then isGz defaults to true and writes to gzipped file)}

\item{faToTwoBit}{path to faToTwoBit program from blat}

\item{tmpDir}{directory to store working files}

\item{...}{additional arguments to run blat}
}
\description{
Make a 2bit file from one set of reads and blat another set against it
}

