% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{cacheOperation}
\alias{cacheOperation}
\title{Cache an operation to disk or load if cache file present}
\usage{
cacheOperation(cacheFile, operation, ..., OVERWRITE = FALSE, VOCAL = TRUE,
  EXCLUDE = NULL)
}
\arguments{
\item{cacheFile}{File location to save data to}

\item{operation}{A function taking ... arguments and returning object to be stored}

\item{...}{Arguments for operation function}

\item{OVERWRITE}{If FALSE throw an error if hash of ... changes from cached values. If TRUE redo operation and overwrite cache without asking}

\item{VOCAL}{If TRUE report on status of caching}

\item{EXCLUDE:}{Vector of names of arguments to exclude from md5 digest comparison (for very large arguments)}
}
\value{
Output from operation function with ... arguments
}
\description{
Cache an operation to disk or load if cache file present
}

