% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{chao}
\alias{chao}
\title{Calculate chao diversity index}
\usage{
chao(counts)
}
\arguments{
\item{counts}{A vector of counts with one entry per "species"}
}
\value{
Chao index
}
\description{
Calculate chao diversity index
}

