% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{check_fix_reaction}
\alias{check_fix_reaction}
\title{Check and fix a reaction}
\usage{
check_fix_reaction(reaction)
}
\arguments{
\item{reaction}{A string representing the reaction.}
}
\value{
The fixed reaction if some fix was needed, or the
reaction received as input if everything is ok.
}
\description{
This function is used for the check and fix (if possible) of a reaction.
It is checked if there is an empty side, if both sides only contains
\code{0}'s, when the relation operator between sides (\code{->}) is missing. It also
fix the reactions that has a \code{0} with other species in the same side,
removing the 0 and warning the user.
}
