% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{check_neuron_circuit_hje}
\alias{check_neuron_circuit_hje}
\title{Checks the circuit structure}
\usage{
check_neuron_circuit_hje(circuit)
}
\arguments{
\item{circuit}{Circuit object to be checked.}
}
\description{
This function checks if the circuit passed as a parameter has the correct
structure.
}
\details{
Currently, it is checking if \code{circuit} is a list and if it has a list
called \code{gates} in it.
}
