% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{check_neuron_gate_binding_hje}
\alias{check_neuron_gate_binding_hje}
\title{Check the gate binding structure}
\usage{
check_neuron_gate_binding_hje(gate_binding)
}
\arguments{
\item{gate_binding}{A list representing a gate binding.}
}
\description{
This function checks if a given object has the structure of a gate binding.
}
\details{
It is checked if the fields \code{gate1_name}, \code{gate2_name}, \code{input_neuron_idx}
\code{neuron_binding_crn} are all set and with the correct types.
}
