% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{check_reaction_4domain}
\alias{check_reaction_4domain}
\title{Check if the reaction is compatible with the ones
supported by \code{\link{react_4domain}()}.}
\usage{
check_reaction_4domain(reaction)
}
\description{
The reactions supported by \code{\link{react_4domain}()} are
only the uni or bimolecular. This function checks if the
\code{reaction} parameter meets this requirement, returning
\code{TRUE} if the reaction is supported, or \code{FALSE}
otherwise.
}
\examples{
DNAr:::check_reaction_4domain('A + B -> C')   # Should return TRUE
DNAr:::check_reaction_4domain('2A -> B')      # Should return TRUE
DNAr:::check_reaction_4domain('2A + B -> C')  # Should return FALSE
}
