% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlat.R
\name{checkBlat}
\alias{checkBlat}
\title{Check if blat is running on port}
\usage{
checkBlat(port)
}
\arguments{
\item{port}{port to check if blat is running on}
}
\value{
TRUE if blat running on port otherwise FALSE
}
\description{
Check if blat is running on port
}
