% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{checkOverlap}
\alias{checkOverlap}
\title{Check overlap between two sets of coordinates}
\usage{
checkOverlap(chrs, starts, ends, tChrs, tStarts, tEnds, tNames, sep = "|")
}
\arguments{
\item{chrs}{Chromosomes of query}

\item{starts}{Start coordinates of query}

\item{ends}{End coordinates of query}

\item{tChrs}{Chromsomes of target}

\item{tStarts}{Start coordinates of target}

\item{tEnds}{End coordinates of target}

\item{tNames}{Target names}

\item{sep}{Matches are pasted together separated by this}
}
\value{
'|' separated vector of tNames within overlap or '' if no overlapping target
}
\description{
Check overlap between two sets of coordinates
}
\examples{
queries<-data.frame(
  'chr'=letters[rep(1:3,each=2)],
  'start'=c(1,1000,-1,1000,-10000,1),
  'end'=c(1010,2000,50,1050,10000,1)
)
targets<-data.frame('chr'=letters[rep(1:2,each=2)],
  'start'=c(1,1111,1,1000),
  'end'=c(1000,9999,2,1000),
  'name'=letters[1:4]
)
checkOverlap(
  queries$chr,queries$start,queries$end,
  targets$chr,targets$start,targets$end,targets$name
)
}

