% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{checkOverlapMulti}
\alias{checkOverlapMulti}
\title{Check overlap between two sets of coordinates (ranges package may be better/quicker option)}
\usage{
checkOverlapMulti(starts, ends, tStarts, tEnds, tNames, qChrom, tChrom,
  vocal = FALSE, ...)
}
\arguments{
\item{starts}{Start coordinates of query}

\item{ends}{End coordinates of query}

\item{tStarts}{Start coordinates of target}

\item{tEnds}{End coordinates of target}

\item{tNames}{Target names}

\item{qChrom}{Chromosomes of the queries}

\item{tChrom}{Chromosomes of the targets}

\item{vocal}{If TRUE report working on each chromosome}

\item{...}{Additional arguments for checkOverlap}
}
\value{
'|' separated vector of tNames within overlap or '' if no overlapping target
}
\description{
Check overlap between two sets of coordinates (ranges package may be better/quicker option)
}

