% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{chooseAtLeastOneFromEach}
\alias{chooseAtLeastOneFromEach}
\title{Calculate the number of combinations possible from nGroups groups with groupsize members and n observations}
\usage{
chooseAtLeastOneFromEach(nGroups, groupSize, n)
}
\arguments{
\item{nGroups}{number of equally sized groups}

\item{groupSize}{size of the equally sized groups}

\item{n}{number of observations pulled from the individuals}
}
\value{
number of combinations containing at least one individual from each group
}
\description{
Calculate the number of combinations possible from nGroups groups with groupsize members and n observations
}

