% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cleanMclapply}
\alias{cleanMclapply}
\title{A function to break an mclapply into parts, save to disk and run independently}
\usage{
cleanMclapply(x, applyFunc, ..., mc.cores = parallel::detectCores(),
  extraCode = "", nSplits = mc.cores, VOCAL = TRUE,
  envir = .GlobalEnv)
}
\arguments{
\item{x}{vector/list to apply over}

\item{applyFunc}{function to apply}

\item{...}{arguments for mclapply}

\item{mc.cores}{number of cores to use}

\item{extraCode}{character vector of setup code (each command self-contained within a cell or concatenated with ;)}

\item{nSplits}{number of splits to make (if > mc.cores then R will restart more frequently and load less data at once)}

\item{VOCAL}{if TRUE then output status information}

\item{envir}{environment to look for variables in}
}
\value{
the concatenated outputs from applyFunc 
cleanMclapply(1:10,2,sqrt)
cleanMclapply(1:10,2,function(x,y)x^y,y=3)
}
\description{
A function to break an mclapply into parts, save to disk and run independently
}
