% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dna.R
\name{codon2aa}
\alias{codon2aa}
\title{Convert codon to amino acid}
\usage{
codon2aa(codons, type = "code", naReplace = "z", warn = TRUE)
}
\arguments{
\item{codons}{Vector of 3 base codons}

\item{type}{Amino acid info to return; code for single letter, name for full name, or abbr for 3-letter abbreviation}

\item{naReplace}{Replace NAs with this string. Anything other than a single character will result in a protein longer than the number of codons}

\item{warn}{Warn if any unknown codons are found. Usually due to ACTG characters}
}
\value{
Vector of amino acids
}
\description{
Convert codon to amino acid
}

