% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{codons2aa}
\alias{codons2aa}
\title{Convert codon to amino acid}
\usage{
codons2aa(codons, type = "code", naReplace = "z", warn = TRUE)
}
\arguments{
\item{codons}{Vector of 3 base codons}

\item{type}{Amino acid info to return; code for single letter, name for full name, or abbr for 3-letter abbreviation}

\item{naReplace}{Replace NAs with this string. Anything other than a single character will result in a protein longer than the number of codons}

\item{warn}{Warn if any unknown codons are found. Usually due to ACTG characters}
}
\value{
Vector of amino acids
}
\description{
Convert codon to amino acid
}
\examples{
codons2aa(c('AUG','GTT','AGG','GTA'))
codons2aa(dna2codons('ATGTTTATTGGG')[[1]])
}
