% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{combine_crns}
\alias{combine_crns}
\title{Combine CRNs into one single larger CRN}
\usage{
combine_crns(crns)
}
\arguments{
\item{crns}{List of CRNs.}
}
\value{
Another CRN representing the input CRNs combined
}
\description{
Use this function to combine multiple crns into a larger crn which
has all the reactions, species and other parameters of all other CRNs.
}
