% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{combine_reaction_parts}
\alias{combine_reaction_parts}
\title{Combines two reaction parts to form a reaction}
\usage{
combine_reaction_parts(left_part, right_part, operator)
}
\arguments{
\item{left_part}{The first part of the reaction}

\item{right_part}{The second part of the reaction}

\item{operator}{A string specifying the relation between the parts.
It will be concatenated between the string parts.}
}
\value{
A string representing the complete construction of a reaction.
}
\description{
This function combines two reaction parts and forms
a complete reaction construction with a reaction operator
which specifies the relation between the two parts.
}
