% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{compare_behaviors_nrmse}
\alias{compare_behaviors_nrmse}
\title{Compare the behavior of two reactions}
\usage{
compare_behaviors_nrmse(bhv_sim, bhv_obs)
}
\arguments{
\item{bhv_sim}{The simulated behavior.}

\item{bhv_obs}{The observed behavior (used in the normalization).}
}
\value{
A data frame with the same columns of the behaviors and one row.
Each value is the NRMSE of that species.
}
\description{
Use this function to compare the behavior of two reactions to see
the similarity between them for each species. The normalized
root-mean-square error (NRMSE) measure (\code{\link{nrmse}()}) is used
to make this comparison.
}
\details{
For convention, one of the behaviors is called
simulated behavior, and the other one is called observed behavior. This
differentiation is important because the order that you pass the
behaviors impacts in the result.

The normalization made by the NRMSE
uses the observed values (\code{bhv_obs} parameter) only, consequently,
\code{compare_behaviors_nrmse(data1, data2)} results in different measures
than \code{compare_behaviors_nrmse(data2, data1)}.
}
