% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{complimentDna}
\alias{complimentDna}
\title{Compliment DNA}
\usage{
complimentDna(dnas, ambigs = TRUE)
}
\arguments{
\item{dnas}{vector of sequences}

\item{ambigs}{if TRUE compliment ambiguous bases}
}
\value{
vector with the DNA sequences complimented
}
\description{
Compliment DNA
}
\examples{
complimentDna(c('CTAG','ATCCAC'))
complimentDna(c('CT[AC]G','ATNRY'))
}

