% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{conservativeBoundary}
\alias{conservativeBoundary}
\title{Get the conservative edge of a confidence interval}
\usage{
conservativeBoundary(boundaries, base = 0)
}
\arguments{
\item{boundaries}{upper and lower values}

\item{base}{value, e.g. 0 or 1}
}
\value{
single numberic. if the boundaries overlap the base then return base otherwise return value closest to base
}
\description{
Get the conservative edge of a confidence interval
}

