% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convertLineToUser}
\alias{convertLineToUser}
\title{Convert from axis line to user coordinates}
\usage{
convertLineToUser(line, axis = 1)
}
\arguments{
\item{line}{line to convert to user coordinates}

\item{axis}{axis to do conversion on (1:4 same as axis, mtext command)}
}
\value{
position in user coordinates
}
\description{
Convert from axis line to user coordinates
}
\examples{
plot(1:10)
points(rep(5,5),convertLineToUser(0:4,1),xpd=NA)
points(convertLineToUser(0:4,2),rep(5,5),xpd=NA)
}
