% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convertUserToLine}
\alias{convertUserToLine}
\title{Convert from user coordinates to axis line}
\usage{
convertUserToLine(usr, axis = 1)
}
\arguments{
\item{usr}{user coordinate to convert to line}

\item{axis}{axis to do conversion on (1:4 same as axis, mtext command)}
}
\value{
axis line position
}
\description{
Convert from user coordinates to axis line
}

