% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{countNmers}
\alias{countNmers}
\title{Count nMers in a string}
\usage{
countNmers(string, k = 10)
}
\arguments{
\item{string}{string to be searched for nmers}

\item{k}{length of nmers}
}
\value{
a sorted named vector giving the identity and counts for nMers
}
\description{
Count nMers in a string
}
\examples{
countNmers('AATTAATT',2)
countNmers('AATTAATT',3)
}
