% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{cumpaste}
\alias{cumpaste}
\title{Cumulative paste convenience function}
\usage{
cumpaste(strs, sep = " ")
}
\arguments{
\item{strs}{a vector of strings}

\item{sep}{a string to be used as the separator in paste}
}
\value{
a vector of strings the same length as strs with the concatenated strings. Note that NAs will be converted to the string "NA" by paste
}
\description{
Cumulative paste convenience function
}
\examples{
cumpaste(letters,sep='')
}
