% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{cutBlatReads}
\alias{cutBlatReads}
\title{Take the output from blat and make continous reads out of it}
\usage{
cutBlatReads(seqs, starts, fillStarts = NULL, fillEnds = NULL,
  gaps = NULL)
}
\arguments{
\item{seqs}{comma separated target sequences from blat}

\item{starts}{comma separated starting location for each piece of sequence 0 indexed}

\item{fillStarts}{start base for each output sequence 0 indexed}

\item{fillEnds}{total length for each output sequence}

\item{gaps}{list of matrices with columns tGaps and qGaps}
}
\value{
vector of sequences
}
\description{
Take the output from blat and make continous reads out of it
}

