% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cv.glm.par}
\alias{cv.glm.par}
\title{Check predictions of glm by cross validation}
\usage{
cv.glm.par(model, data = eval(modelCall$data), K = nrow(data),
  nCores = 1, subsets = NULL, vocal = FALSE)
}
\arguments{
\item{model}{a glm to cross validate}

\item{data}{data for the model}

\item{K}{number of pieces to split data into}

\item{nCores}{number of cores to use}

\item{subsets}{predefined subset of same length as data specifying groupings}

\item{vocal}{echo progress indicator}
}
\value{
a dataframe giving predictions for each data point with columns prediction and subsetId
}
\description{
Check predictions of glm by cross validation
}
\examples{
z<-data.frame('x'=1:11,y=c(rnorm(10,1:10),100))
cv.glm.par(glm(y~x,data=z))
}
