% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{define_neuron_gate_binding_hje}
\alias{define_neuron_gate_binding_hje}
\title{Get the binding between two neuron gates}
\usage{
define_neuron_gate_binding_hje(gate1, gate2, input_neuron_idx, ...)
}
\arguments{
\item{gate1, gate2}{The gates which will be binded}

\item{input_neuron_idx}{The index of the input neuron of \code{gate2} which
will be binded.}

\item{...}{The parameters \code{enzyme_config}, \code{ci} and optionally
\code{bind_inhibitory}, which will be passed to
\code{\link{get_neuron_binding_hje}()}}
}
\value{
The \code{gate2} received as input but with the specified input neuron
binded to the \code{gate1} output.
}
\description{
This function returns a modified gate of \code{gate2} input. this modified
version has one of its input neurons binded to the output neuron of \code{gate2}.
}
\details{
The binding is made using the function
\code{\link{add_gate_binding_on_neuron_hje}()}. The binding is a list with
the following structure:
\itemize{
\item \code{gate1_name}, \code{gate2_name}: = gate 1 and 2 names, respectively;
\item \code{input_neuron_idx}: the parameter \code{input_neuron_idx},
\item \code{neuron_binding_crn}: binding CRN returned by
\code{\link{get_neuron_binding_hje}()}
}
}
\seealso{
\code{\link{get_neuron_binding_hje}()} for details about the parameter
that it s expecting.
}
