% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{degap}
\alias{degap}
\title{Remove gap characters from DNA sequences}
\usage{
degap(seq, gaps = c("*", "-", "."))
}
\arguments{
\item{seq}{vector of gapped DNA sequences}

\item{gaps}{vector of characters to be considered gaps}
}
\value{
vector of DNA sequences with gaps removed
}
\description{
Remove gap characters from DNA sequences
}

