% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{dna2codons}
\alias{dna2codons}
\title{Convert dna string into seperate codons}
\usage{
dna2codons(dna, frame = 0)
}
\arguments{
\item{dna}{Single string of DNA}

\item{frame}{Starting frame for codons (0=start on first base, 1=on second, 2=on third)}
}
\value{
Vector of 3 base codons
}
\description{
Convert dna string into seperate codons
}

