% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{dnaPos2aa}
\alias{dnaPos2aa}
\title{Find a single codon at a given position in dna}
\usage{
dnaPos2aa(dna, pos, start = 1, end = nchar(dna), frame = 0,
  strand = "+", refStart = 1, debug = FALSE)
}
\arguments{
\item{dna}{A string of DNA/RNA}

\item{start}{Start coordinate of exon}

\item{end}{End coordinate of exon}

\item{frame}{Starting frame (0=start on first base, 1=on second, 2=on third)}

\item{strand}{Strand of dna (i.e. revcomp the dna first if '-')}

\item{refStart}{Start coordinate of string}

\item{debug}{If TRUE print debug info}
}
\value{
Amino acid code or NA if outside exon or outside string
}
\description{
Find a single codon at a given position in dna
}

