% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsd.R
\name{dsd_4d_module_str}
\alias{dsd_4d_module_str}
\title{Returns a string with the 4-domain dsd module}
\usage{
dsd_4d_module_str(reaction, species_domains, k_idx)
}
\arguments{
\item{reaction}{The reaction for which a dsd module will be
instantiated}

\item{species_domains}{A list of lists in which there is a list of domains
for each species.}

\item{k_idx}{The index of k, used to name the k variable.}
}
\value{
A string representing the dsd module instantiated.
}
\description{
Use this function to get a string representing a 4-domain
dsd module for your reaction.
}
