% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsd.R
\name{dsd_4d_modules}
\alias{dsd_4d_modules}
\title{Get the available 4-domain modules for DSD script}
\usage{
dsd_4d_modules()
}
\value{
The matrix of functions which can be accesses with [[i,j]]

[[i,j]: R:[i,j
}
\description{
This function returns a matrix with all available 4-domain modules
for the DSD script. The matrix is organized based on the reaction
stoichiometry.
}
\details{
If a reaction has 2 reactants and 1 product,
will be in the position 3,2. The formation reactions are in the
line 1, and the degradation reactions are in the column 1 ([[,1]]).

The position [[1,1]] is filled with \code{NA} because it represents an
invalid reaction (\code{0 -> 0}).

[[,1]: R:[,1
[[1,1]: R:[1,1
}
