% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{escapeRegexBracketChars}
\alias{escapeRegexBracketChars}
\title{Convenience function to escape characters for insertion into regex brackets}
\usage{
escapeRegexBracketChars(regexChars, escapeChars = c("^", "-", "[", "]",
  "\\\\"))
}
\arguments{
\item{regexChars}{a vector of single characters to escape if in escapeChars}

\item{escapeChars}{a vector of which characters to escape}
}
\value{
a vector of escaped characters
}
\description{
Note to escape a '\code{]}' using this you'll have to use \code{perl=TRUE} in the downstream function
}
\examples{
escapeRegexBracketChars(c(']','\\\\','-','A','B'))
}
