% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{eval_derivative}
\alias{eval_derivative}
\title{Evaluate an expression a derivative returned by
\code{\link{analyze_behavior}()}}
\usage{
eval_derivative(derivative)
}
\arguments{
\item{derivative}{Derivative with concentration values returned by
\code{\link{analyze_behavior}()}.}
}
\value{
A numeric value representing the result of the derivative.
}
\description{
If \code{\link{analyze_behavior}()} was used with a \code{behavior} and
\code{time_point}, you can use this function to calculate the result of
the derivative.
}
