% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{eval_derivative_part}
\alias{eval_derivative_part}
\title{Evaluate subexpressions of a derivative returned by
\code{\link{analyze_behavior}()}}
\usage{
eval_derivative_part(derivative)
}
\arguments{
\item{derivative}{Derivative with concentration values returned by
\code{\link{analyze_behavior}()}.}
}
\value{
A numeric named vector value representing the result of each part
of the the derivative. The names of the results are the evaluated
subexpressions.
}
\description{
This functions works like \code{\link{eval_derivative}()}, but instead of
evaluating the entire derivative, it will evaluate parts of it
(delimited by \code{()}).
}
