% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{expand_species}
\alias{expand_species}
\title{Function to expand species according to their stoichiometry}
\usage{
expand_species(species, stoichiometry)
}
\arguments{
\item{species}{The list of species;}

\item{stoichiometry}{A function that calculates the stoichiometry
of the species.}
}
\value{
An expanded list of species.
}
\description{
This function expand a list of species according to the stoichiometry
of the species. E.g.: given a list of species \code{('A', 'B')} and
a stoichiometry of \code{A = 1} and \code{B = 2}, the resulting list
will be a list \code{('A', 'B', 'B')}.
}
