% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{expandAmbiguous}
\alias{expandAmbiguous}
\title{Convert ambiguous dna to all possible sequences}
\usage{
expandAmbiguous(dna)
}
\arguments{
\item{dna}{vector dna containing ambiguous bases}
}
\value{
list with each entry containing all combinations of ambiguous bases for that entry of the dna vector
}
\description{
Convert ambiguous dna to all possible sequences
}
\examples{
expandAmbiguous(c('AACACANAT',"ACCHB"))
}
\references{
\url{https://en.wikipedia.org/wiki/Nucleic_acid_notation}
}

