% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fillDown}
\alias{fillDown}
\title{Convenience function to fill values down a column}
\usage{
fillDown(x, emptyStrings = c(NA, ""))
}
\arguments{
\item{x}{a vector to be filled}

\item{emptyStrings}{a vector of strings to be considered empty. Defaults to empty string and NA.}
}
\value{
a vector of the same length as x with empty entries filled
}
\description{
Takes a vector of values and fills any empty strings or NAs with the last non-empty value.
}
\examples{
fillDown(c(1:5,NA,NA,6,NA,7))
fillDown(c('a','c','d','Z',NA),'Z')
}

