% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fillList}
\alias{fillList}
\title{Loop through list, get unique names and make sure every element has those names}
\usage{
fillList(x, namesList = unique(unlist(lapply(x, names))), fill = NA)
}
\arguments{
\item{x}{list to loop through}

\item{namesList}{names to make sure every element has (and delete extras)}

\item{fill}{value to insert in missing elements}
}
\value{
list with missing elements filled in and elements not in namesList removed
}
\description{
Loop through list, get unique names and make sure every element has those names
}
\examples{
fillList(list(c('a'=1,'b'=2),c('c'=3)))
fillList(list(c('a'=1,'b'=2),c('c'=3)),namesList='a')
}

