% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{fillZeros}
\alias{fillZeros}
\title{Fill missing position with zeros}
\usage{
fillZeros(pos, data, fillValue = 0)
}
\arguments{
\item{pos}{a vector integer position where any missing integers will be filled in}

\item{data}{a data.frame of data with missing rows to be filled with zeros}

\item{fillValue}{value to fill in missing data with}
}
\value{
a 2 element list with a vector of positions and a data.frame of filled data
}
\description{
Fill missing position with zeros
}
\examples{
fillZeros(c(1:3,6:7,9),data.frame('count'=1:6))
}

