% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/454.R
\name{flow2seq}
\alias{flow2seq}
\title{Convert 454 flows to DNA sequences}
\usage{
flow2seq(flow, flowOrder = c("T", "A", "C", "G"))
}
\arguments{
\item{flow}{vector of flow values (e.g. produced by seq2flow)}

\item{flowOrder}{order of nucleotide flows}
}
\value{
DNA sequence corresponding to the flowgram
}
\description{
Convert 454 flows to DNA sequences
}

