% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fte_theme}
\alias{fte_theme}
\title{\code{ggplot2} theme developed by Max Woolf.}
\usage{
fte_theme()
}
\description{
This is a simple theme for \code{ggplot2} package designed by Max Woolf.
}
\references{
\itemize{
\item \href{http://minimaxir.com/2015/02/ggplot-tutorial/}{Website with the theme.}
}
}
