% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dna.R
\name{gap2NoGap}
\alias{gap2NoGap}
\title{Convert gapped coordinates to what the coordinates would be without gaps}
\usage{
gap2NoGap(gapSeq, coords)
}
\arguments{
\item{gapSeq}{the reference sequence containing gaps}

\item{coords}{coordinates on the gapped gapSeq to be converted into equivalent nongap coordinatess}
}
\value{
equivalent gapped coordinates
}
\description{
Convert gapped coordinates to what the coordinates would be without gaps
}

