% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{gcPercent}
\alias{gcPercent}
\title{Calculate GC percentage for strings}
\usage{
gcPercent(seqs, chars = c("C", "G"))
}
\arguments{
\item{seqs}{sequences to get GC content of}

\item{chars}{characters to count (G,C by default)}
}
\value{
proportion of GC in sequence
}
\description{
Calculate GC percentage for strings
}
\examples{
gcPercent(c('AAATA','AATTC','GCGCTT'))
gcPercent(c('12345','22234','01011101'),'1')
}
