% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{generateFasta}
\alias{generateFasta}
\title{Generate fake name and sequence data}
\usage{
generateFasta(nSeq = 10000, nChar = 100:1000, bases = c("A", "C", "T",
  "G", "-", "N"), generateQuals = FALSE, qualRange = 1:40, baseQual = 33)
}
\arguments{
\item{nSeq}{number of sequences to generate}

\item{nChar}{possible lengths of sequences}

\item{bases}{possible bases}

\item{generateQuals}{if TRUE generate qualities for each sequence}

\item{qualRange}{The range of possible qualities}

\item{baseQual}{integer to subtract from quality characters to convert into quality space}
}
\value{
data.frame with nSeq rows and columns name and seq (and qual if generateQuals is TRUE)
}
\description{
Generate fake name and sequence data
}
\examples{
generateFasta(10,10:20)
generateFasta(10,10:20,generateQuals=TRUE)
}
