% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{get_circuit_from_neuron_gates_hje}
\alias{get_circuit_from_neuron_gates_hje}
\title{Get a circuit from a list of gates}
\usage{
get_circuit_from_neuron_gates_hje(gates)
}
\arguments{
\item{gates}{A list of gates.}
}
\value{
A circuit, which is a list with the following fields in it:
\itemize{
\item \code{gates}: The list of gates.
}
}
\description{
This function checks if the gates and gate bindings are correctly defined
(using \code{\link{check_neuron_gate_hje}()} and \code{\link{check_neuron_gate_hje}()}
functions). Then, it structures all gates into one single list called
circuit. The circuit represents all the gates and their interactions,
defining one single object containing all the gate network.
}
