% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{get_crn_from_neuron_circuit_hje}
\alias{get_crn_from_neuron_circuit_hje}
\title{Get a CRN from a circuit}
\usage{
get_crn_from_neuron_circuit_hje(circuit)
}
\arguments{
\item{circuit}{The circuit (returned by
\code{\link{get_circuit_from_neuron_gates_hje}()})}
}
\value{
A list specifying the CRN.
}
\description{
This function returns a CRN given a circuit as a parameter.
}
\details{
In this function, the gate bindings are correctly set, updating
the neuron which will receive the bindings as input. The bindings
within the gates are assumed to be already setted.
}
