% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{get_crn_from_neuron_gate_hje}
\alias{get_crn_from_neuron_gate_hje}
\title{Get a CRN given a gate}
\usage{
get_crn_from_neuron_gate_hje(gate)
}
\arguments{
\item{gate}{List representing the gate (returned by
\code{\link{get_neuron_generic_gate_hje}()}, for example).}
}
\value{
A CRN representing the \code{gate}.
}
\description{
This function returns a list specifying a CRN given a list
specifying a gate as a parameter, combining all the CRNs of the gate
into one single CRN.
}
