% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsd.R
\name{get_dsd_4domain_modules_str}
\alias{get_dsd_4domain_modules_str}
\title{Get the 4-domain modules for the DSD script.}
\usage{
get_dsd_4domain_modules_str()
}
\value{
String with the 4-domain modules.
}
\description{
Use this function to get a string with all 4-domain
modules for the DSD script.
}
