% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{get_dsd_buff_str}
\alias{get_dsd_buff_str}
\title{Instantiate a buffer module in the DSD script}
\usage{
get_dsd_buff_str(qs, qmax, Cmax, Cii, d, domains)
}
\arguments{
\item{qs}{String representing a variable name for the \code{qs}
parameter of the module signature.}

\item{qmax}{String representing a variable name for the \code{qmax}
parameter of the module signature.}

\item{Cmax}{String representing a variable name for the \code{Cmax}
parameter of the module signature.}

\item{Cii}{String representing a variable name for the \code{Cii}
parameter of the module signature.}

\item{d}{String representing a variable name for the \code{d}
parameter of the module signature.}

\item{domains}{Vector of strings representing the domains of
the input species.}
}
\value{
A string representing the instantiation of a \code{Buff()} module.
}
\description{
This function returns a string representing an addition of
a buffer reaction module in the DSD script. It creates a \code{Buff()}
module in the script, replacing all the parameter strings by the ones
specified in this function.
}
