% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{get_dsd_def_str}
\alias{get_dsd_def_str}
\title{Get a definition string for the DSD script}
\usage{
get_dsd_def_str(key, val)
}
\arguments{
\item{key}{String representing the key of the definition.}

\item{val}{String representing the value of the definition.}
}
\value{
The definition string.
}
\description{
This function returns a string representing a definition in the
DSD script.
}
\details{
A definition is made by the following construction:
\code{def *key* = *value*}. passing a key and value to this function,
it returns a definition construction replacing the \emph{key} and \emph{value}
by the parameters \code{key} and \code{value} specified.
}
