% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{get_dsd_header_str}
\alias{get_dsd_header_str}
\title{Get the DSD script header}
\usage{
get_dsd_header_str(time, species_to_plot)
}
\arguments{
\item{time}{The time sequence that simulation will occour.
this parameter can be exactly the same used in
the parameter \code{t} of the \code{\link{react}()}
or \code{\link{react_4domain}()} functions.}

\item{species_to_plot}{A vector with the species names to be plotted.}
}
\value{
A string with the DSD script header.
}
\description{
Use this function to get the header of the DSD script.
This header comes with simulation time, ODE solver,
the concentration unit measure, type of compilation and
the species to be plotted.
}
