% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{get_dsd_species_str}
\alias{get_dsd_species_str}
\title{Get the DSD definition of a species}
\usage{
get_dsd_species_str(species_name, domains)
}
\arguments{
\item{species_name}{The species name.}

\item{domains}{A vector of four string representing the
domain names of this species.}
}
\value{
A string with the species definition.
}
\description{
Use this function to get a string representing the definition
of a species.
}
