% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{get_neuron_AND_gate_hje}
\alias{get_neuron_AND_gate_hje}
\title{Get a AND neuron gate}
\usage{
get_neuron_AND_gate_hje(gate_name, input_cis, binding_inhibit = NULL)
}
\arguments{
\item{gate_name}{The name of the gate}

\item{input_cis}{A numeric vector representing the initial
concentration of the inputs.}

\item{binding_inhibit}{Boolean vector where each elements specifies
whether the input with the same index will
be an inhibitory binding or not. If no vector
is passed, no binding will be inhibited. This
parameter is passed to
\code{\link{get_neuron_generic_gate_hje}()}.}
}
\value{
A list representing the gate. To transform this representation into
a CRN (to simulate with \code{\link{react}()}, for example) you have
to use the function \code{\link{get_crn_from_neuron_gate_hje}()}.
}
\description{
This function returns an AND gate using neurons based on McCullogh-Pitts
model (neurons returned by \code{\link{get_neuron_hje}()}). The two input neurons
have the name \code{i1} and \code{i2}, and the output neuron has the \code{o} name. This
(with the gate name) determines the species names created for the CRNs
that represent them.
}
\examples{
g         <- DNAr::get_neuron_AND_gate_hje('_AND', c(0.5, 1.5))
and_crn   <- DNAr::get_crn_from_neuron_gate_hje(g)
and_crn$t <- seq(0, 1, length.out = 100)
b         <- do.call(DNAr::react, and_crn)
DNAr::plot_behavior(b, c('A_ANDo'))
}
