% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{get_neuron_gate_binding_hje}
\alias{get_neuron_gate_binding_hje}
\title{Get the binding between two neuron gates}
\usage{
get_neuron_gate_binding_hje(gate1, gate2, input_neuron_idx, ...)
}
\arguments{
\item{gate1, gate2}{The gates which will be binded}

\item{input_neuron_idx}{The index of the input neuron of \code{gate2} which
will be binded.}

\item{...}{The parameters \code{enzyme_config}, \code{ci} and optionally
\code{bind_inhibitory}, which will be passed to
\code{\link{get_neuron_binding_hje}()}}
}
\value{
Get the binding CRN between gates.
}
\description{
This function returns a CRN which represents a binding between the
output neuron of one gate (\code{gate1}) to one of the input neurons of
the other gate (\code{gate2}).
}
\seealso{
\code{\link{get_neuron_binding_hje}()} for details about the parameter
that it s expecting.
}
