% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{get_neuron_generic_gate_hje}
\alias{get_neuron_generic_gate_hje}
\title{Get a generic neuron gate with n inputs and 1 output}
\usage{
get_neuron_generic_gate_hje(input_neuron_names, output_neuron_name,
  input_neuron_cis, binding_enzyme_configs, binding_cis,
  binding_inhibit = NULL)
}
\arguments{
\item{output_neuron_name}{A string being the output neuron name}

\item{input_neuron_cis}{A vector of number representing the
initial concentration of each neuron}

\item{binding_enzyme_configs}{A list where each element represents
a enzyme configuration. Each configuration
is a list with the following parameters:
\itemize{
\item \code{ci}: Initial concentration of the
enzyme
\item \code{k}: Vector with the rate constants
of forward and backward reactions,
respectively.
}}

\item{binding_cis}{Numeric vector with the initial concentration
of the bindings}

\item{binding_inhibit}{Boolean vector where each elements specifies
whether the binding with the same index will
be an inhibitory binding or not. If no vector
is passed, no binding will be inhibited.}

\item{input_neuron_name}{A list of string. Each string specifies}
}
\value{
A list representing the gate. This list comes with three parameters:
\itemize{
\item \code{output_neuron_crn}: A neuron CRN returned by
\code{\link{get_neuron_hje}()};
\item \code{input_neuron_crns}: A list of neuron CRNs also returned by
\code{\link{get_neuron_hje}()};
\item \code{binding_crns}: A list of binding CRNs returned by
\code{\link{get_neuron_binding_hje}()}, which are the binding
of each input neuron with the output neuron.
}
}
\description{
This function returns a generic neuron gate. Generic because the
number of inputs of this gate will be specified as an argument
(length of \code{input_neuron_names}).
}
\details{
A neuron gate is a gate made with neurons, more specifically, 1 + n neurons,
where n is the number of inputs. Each input is a binary neuron that passes
its output to the last neuron, the output neuron. This neuron has as input
the sum of the output of all input neurons.
}
\seealso{
\code{\link{react}()} for more about the neuron behavior.
}
