% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{get_neuron_hje}
\alias{get_neuron_hje}
\title{Get a neuron based on Hjelmfelt A. et al. \code{[1]} approach}
\usage{
get_neuron_hje(name, input_ci)
}
\arguments{
\item{name}{The neuron name. This will be used for the species names.}

\item{input_ci}{A numeric value specifying the initial concentration of
the input}
}
\value{
A CRN representing the neuron.
}
\description{
This function returns a neuron based on the model described by
Hjelmfelt A. et al. \code{[1]}.
}
\references{
\itemize{
\item \code{[1]} \insertRef{hjelmfelt1991chemical}{DNAr}
}
}
