% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{get_neuron_majority_gate}
\alias{get_neuron_majority_gate}
\title{Get a majority gate}
\usage{
get_neuron_majority_gate(gate_name, n_inputs, input_cis,
  binding_inhibit = NULL)
}
\arguments{
\item{gate_name}{The name of the gate.}

\item{n_inputs}{The number of inputs.}

\item{input_cis}{A numeric vector representing the initial
concentration of the inputs.}

\item{binding_inhibit}{Boolean vector where each elements specifies
whether the input with the same index will
be an inhibitory binding or not. If no vector
is passed, no binding will be inhibited. This
parameter is passed to
\code{\link{get_neuron_generic_gate_hje}()}.}
}
\value{
A list representing the gate. To transform this representation into
a CRN (to simulate with \code{\link{react}()}, for example) you have
to use the function \code{\link{get_crn_from_neuron_gate_hje}()}.
}
\description{
It returns a majority gate using neurons based on McCullogh-Pitts
model (neurons returned by \code{\link{get_neuron_hje}()}). It supports an
arbitrary number of inputs (although it must be odd).
}
