% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_onespecies_count}
\alias{get_onespecies_count}
\title{Get the count of occurrences of a given species in a reaction part.}
\usage{
get_onespecies_count(one_species, reaction_part)
}
\description{
With the reaction part 'A + B ', for instance, and \code{one_species}
begin equal to 'A', this function would return 1. On the case of '2A ' it
would return 2.
}
