% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_products}
\alias{get_products}
\title{Get the products of a given reaction}
\usage{
get_products(reaction)
}
\description{
This function returns the products of a reactions,
removing their stoichiometry. If the reaction is of the type
'A -> 0', an empty string '' is returned.
}
\examples{
DNAr:::get_products('A + B -> C')   # Returns c('C')
DNAr:::get_products('2A -> B + C')  # Returns c('B', 'C')
DNAr:::get_products('A -> 0')       # Returns c('')
}
