% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{get_reactants}
\alias{get_reactants}
\title{Get the reactants of a given reaction}
\usage{
get_reactants(reaction)
}
\description{
This function returns the reactants of a reactions,
removing their stoichiometry.
}
\examples{
DNAr:::get_reactants('A + B -> C')  # Returns c('A', 'B')
DNAr:::get_reactants('2A -> B')     # Returns c('A')
}
