% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{get_stoichiometry_all}
\alias{get_stoichiometry_all}
\title{Get the stoichiometry of a reaction}
\usage{
get_stoichiometry_all(reaction)
}
\value{
a list with left_sto and right_sto being the stoichiometry
of the left and right part respectively.
}
\description{
Use this function to get the stoichiometry of the left and
right part of a reaction.
}
\examples{
# Returns list(left_sto = 2, right_sto = 1)
DNAr:::get_stoichiometry_all('A + B -> C')
# Returns list(left_sto = 1, right_sto = 2)
DNAr:::get_stoichiometry_all('A -> 2B')
}
