% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{get_stoichiometry_onespecies}
\alias{get_stoichiometry_onespecies}
\title{Get the stoichiometry of a specific species in a reaction.}
\usage{
get_stoichiometry_onespecies(one_species, reaction)
}
\value{
A list with \code{left_sto} being the stoichiometry of a
species in the left side of a reaction, and \code{right_sto} being
the same but for the right side of the reaction.
}
\description{
This function uses \code{\link{get_onespecies_count}} in the left
and right part of a reaction to get the stoichiometry of a species
in a reaction.
}
\examples{
# It should return list(left_sto = 1, right_sto = 2)
get_stoichiometry_onespecies('A', 'A + B -> 2A')
# It should return list(left_sto = 0, right_sto = 1)
get_stoichiometry_onespecies('A', 'B -> A + B')
}
