% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_stoichiometry_part}
\alias{get_stoichiometry_part}
\title{It returns the stoichiometry of part of a reaction.}
\usage{
get_stoichiometry_part(reaction_part)
}
\description{
This function is used when you want to know the count of molecules
in part of a reaction.
}
\examples{
DNAr:::get_stoichiometry_part('A + B ')  # It should return 2
DNAr:::get_stoichiometry_part('2A ')     # It should return 2
DNAr:::get_stoichiometry_part(' C')      # It should return 1
}
