% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{gini}
\alias{gini}
\title{Calculate Gini index}
\usage{
gini(x)
}
\arguments{
\item{x}{a vector of numerical values}
}
\value{
a single numeric giving the Gini index
}
\description{
Calculate Gini index
}
\examples{
gini(1:10)
gini(rep(1,100))
gini(c(rep(0,100),1))
}
\references{
\url{https://en.wikipedia.org/wiki/Gini_coefficient}
}
