% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dna.R
\name{hashString}
\alias{hashString}
\title{Generate hases from a string}
\usage{
hashString(string, hashSize, everyBase = 1, start = 1)
}
\arguments{
\item{string}{String to be hashed}

\item{hashSize}{Size of hashed strings}

\item{everyBase}{Generate a hash every X strings}

\item{start}{Output labels start from e.g. label starting from 1000}
}
\value{
Output from operation function with ... arguments
}
\description{
Generate hases from a string
}

