% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{highlightDifferences}
\alias{highlightDifferences}
\title{Switch the case of positions in seq1 that mismatch seq2}
\usage{
highlightDifferences(seq1, seq2)
}
\arguments{
\item{seq1}{sequence to highlight differences in}

\item{seq2}{sequence to compare to}
}
\value{
seq1 with bases that differ from seq2 having their case toggled (e.g. upper to lower case)
highlightDifferences('aaccat','aacgat')
highlightDifferences('TATATTCTTT','TATATCCTTT')
}
\description{
Switch the case of positions in seq1 that mismatch seq2
}

