% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{highlightString}
\alias{highlightString}
\title{Find pattern in strings and flip case}
\usage{
highlightString(pattern, strings)
}
\arguments{
\item{pattern}{Pattern to look for in strings}

\item{strings}{Strings in which to look for pattern}
}
\value{
Strings with case of any occurences of pattern swapped
}
\description{
Use the case of a string to highlight a pattern. For example, highlighting AC in TTTTGTACAAATC as TTTTGTacAAATC. Note that it will not find overlapping patterns e.g. looking for AGA in AGAGA will only find the first AGA. Non-letter characters can be included in the pattern but their case will remain unchanged.
}
\examples{
highlightString('AGA',c('TTAGATTAGA','AGAGA'))
highlightString('aga',tolower(c('TTAGATTAGA','AGAGA')))
}
