% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dna.R
\name{index2range}
\alias{index2range}
\title{Find covered ranges with numerical index data}
\usage{
index2range(index)
}
\arguments{
\item{index}{numeric indices indicating positions of interest}
}
\value{
data.frame with rows for each contiguous region of interest with columns start and end of regions
}
\description{
Find covered ranges with numerical index data
}

