% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/454.R
\name{indexFlow}
\alias{indexFlow}
\title{Find flow corresponding to a given base position}
\usage{
indexFlow(flow, coords)
}
\arguments{
\item{flow}{vector of flow values (e.g. produced by seq2flow)}

\item{coords}{bp coordinates of desired positions}
}
\value{
flow number for each coord
}
\description{
Find flow corresponding to a given base position
}

