% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{indexMatrix}
\alias{indexMatrix}
\title{Convenience function for selecting multiple elements from a matrix by x,y position}
\usage{
indexMatrix(x, y, mat, returnIndex = FALSE)
}
\arguments{
\item{x}{X coordinate of matrix}

\item{y}{Y coorinate of matrix}

\item{mat}{Matrix of interest}

\item{returnIndex}{If TRUE return the one dimensional index for the items otherwise return the selected elements of the matrix}
}
\value{
A vector of indices if returnIndex is TRUE or a vector of the selected matrix elements
}
\description{
Convenience function for selecting multiple elements from a matrix by x,y position
}

