% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{intsToQual}
\alias{intsToQual}
\title{Convert fastq chars to integer quals}
\usage{
intsToQual(ints, baseQual = 33)
}
\arguments{
\item{ints}{A vector of integers convert to quals}

\item{baseQual}{integer to subtract from quality characters to convert into quality space}
}
\value{
a string with one character per quality
}
\description{
Convert fastq chars to integer quals
}
\examples{
intsToQual(1:40)
}
