% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{is_bimolecular}
\alias{is_bimolecular}
\title{Check if a reactions bimolecular.}
\usage{
is_bimolecular(react_str)
}
\description{
Give a reaction, this functions checks if it is
a bimolecular reaction.
}
\examples{
DNAr:::is_bimolecular('2A -> B')     # Should return TRUE
DNAr:::is_bimolecular('A + B -> C')  # Should return TRUE
DNAr:::is_bimolecular('A -> B')      # Should return FALSE
}
