% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{is_degradation}
\alias{is_degradation}
\title{Check if a reaction is of the degradation type}
\usage{
is_degradation(react_str)
}
\arguments{
\item{react_str}{A string representing the reaction}
}
\value{
TRUE if the reaction is of degradation.
}
\description{
Given a reaction string, this function checks if this
reaction is a degradation reaction.
}
