% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{is_formation}
\alias{is_formation}
\title{Check if a reaction is of the formation type}
\usage{
is_formation(react_str)
}
\arguments{
\item{react_str}{A string representing the reaction}
}
\value{
TRUE if the reaction is of formation.
}
\description{
Given a reaction string, this function checks if this
reaction is a formation reaction.
}
