% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{is_unimolecular}
\alias{is_unimolecular}
\title{Check if a reaction is unimolecular.}
\usage{
is_unimolecular(react_str)
}
\description{
Given a reaction, this functions checks if it is
a unimolecular reaction.
}
\examples{
DNAr:::is_bimolecular('2A -> B')     # Should return FALSE
DNAr:::is_bimolecular('A + B -> C')  # Should return FALSE
DNAr:::is_bimolecular('A -> B')      # Should return TRUE
DNAr:::is_bimolecular('0 -> A')      # Should return FALSE
}
