% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{isempty_part}
\alias{isempty_part}
\title{Check if part of a reaction is equal to 0.}
\usage{
isempty_part(react_part)
}
\description{
This function is useful when you have a reaction and
you want to check if it is something like 'A -> 0' (something to nothing).
To do this you get the second part of the reaction with
\code{\link{get_second_part}()} and then use this function with the second
part as the parameter.
}
\examples{
sp <- DNAr:::get_second_part('A -> 0')
DNAr:::isempty_part(sp)  # Should return TRUE
}
