% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{isError}
\alias{isError}
\title{Convenience function to check for errors}
\usage{
isError(x)
}
\arguments{
\item{x}{vector or list to check for errors}
}
\value{
logical vector of length(x) specifying if each element of x was an error
}
\description{
Convenience function to check for errors
}

