% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{jn}
\alias{jn}
\title{Helper function to concatenate strings}
\usage{
jn(...)
}
\arguments{
\item{...}{The strings to be concatenated.}
}
\value{
A string with all the input strings concatenated.
}
\description{
this function should be used for string concatenation without
space between the strings as the default behavior.
}
