% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lagNA}
\alias{lagNA}
\title{Convenience function for lagging with NA}
\usage{
lagNA(x, lag = 1, fill = NA)
}
\arguments{
\item{x}{a vector to be lagged}

\item{lag}{an integer for how much to lag x. lag greater than zero moves values left. lag less than zero moves values right}

\item{fill}{the value to fill at the start or end  of the output lagged vector}
}
\value{
a vector the same length as x
}
\description{
Convenience function for lagging with NA
}

