% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlat.R
\name{liftCoords}
\alias{liftCoords}
\title{Run liftover to convert coordinates from one genome version to another}
\usage{
liftCoords(chr, start, end, strand, chainFile, liftoverBin = "liftOver",
  vocal = FALSE, removeNAs = TRUE)
}
\arguments{
\item{chr}{vector of chromosomes}

\item{start}{vector of start coordinate 1-based}

\item{end}{vector of end coordinate 1-based}

\item{strand}{vector of strands}

\item{chainFile}{location of chain file for liftover}

\item{liftoverBin}{location of the liftOver executable}

\item{vocal}{if TRUE then message extra information}

\item{removeNAs}{if TRUE then remove positions that did could not be lifted over}
}
\value{
four column data frame of chr, start, end and strand of coordinates
}
\description{
Run liftover to convert coordinates from one genome version to another
}

