% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{load_behavior_csv}
\alias{load_behavior_csv}
\title{Load a reaction behavior}
\usage{
load_behavior_csv(filename)
}
\arguments{
\item{filename}{The name of the file that will be loaded. You don't need
to specify the extension of the file, the \code{.csv} extension
will be added automatically at the end of the name
specified.}
}
\value{
A data frame with the reaction behavior.
}
\description{
This function loads a reaction behavior saved by
\code{\link{save_behavior_csv}()}.
}
