% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logAxis}
\alias{logAxis}
\title{Add a pretty log axis to a plot}
\usage{
logAxis(side = 2, exponent = TRUE, addExtra = !exponent,
  minorTcl = -0.2, axisMin = -Inf, ...)
}
\arguments{
\item{side}{an integer specifying which side of the plot the axis is to be drawn on.  The axis is placed as follows: 1=below, 2=left, 3=above and 4=right.}

\item{exponent}{If true label sides using exponents of 10. Otherwise just use normal numbers}

\item{addExtra}{If true add additional labels at 5 or 2 if there are few labels}

\item{minorTcl}{Length of the minor unlabelled ticks. NA to suppress}

\item{axisMin}{The minimum to extend ticks to}

\item{...}{Additional arguments to \code{axis}}
}
\value{
An invisible list containing the minor and major axis label positions
}
\description{
Adds a log axis to the current plot with \code{log='x'} or \code{log='y'} with ticks and labels specified with powers of 10
}
\examples{
plot(1:1000,yaxt='n',log='xy',xaxt='n')
logAxis(las=1)
plot(1:1000,yaxt='n',log='xy',xaxt='n')
logAxis(1,las=1,exponent=FALSE)
}

