% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/454.R
\name{makeSeperateSffs}
\alias{makeSeperateSffs}
\title{Break 454 file(s) into separate files for each sample}
\usage{
makeSeperateSffs(names, samples, barcodes, outDir, sffDir, baseName = "reads",
  sffBinDir = "", vocal = TRUE)
}
\arguments{
\item{names}{names of reads e.g. >DRDR12A125}

\item{samples}{sample ID for each above read}

\item{barcodes}{list of vectors of barcodes or barcode/primers with each entry indexed by sample}

\item{outDir}{directory to put seperate sffs}

\item{sffDir}{directory to look for sffs}

\item{baseName}{prepend to sample names}

\item{sffBinDir}{location of sfffile binary}

\item{vocal}{if TRUE output status messages}
}
\description{
Break 454 file(s) into separate files for each sample
}

