% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mostAbundant}
\alias{mostAbundant}
\title{Convenience function for picking first most abundant of a set of values}
\usage{
mostAbundant(values)
}
\arguments{
\item{values}{A vector of items}
}
\value{
First most abundant item as a string
}
\description{
Convenience function for picking first most abundant of a set of values
}

