% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{movingStat}
\alias{movingStat}
\title{Calculating a moving average/max/min statistic over a vector}
\usage{
movingStat(vec, statFunc = max, spacer = 2)
}
\arguments{
\item{vec}{to average over}

\item{statFunc}{function to use}

\item{spacer}{how many neighbors to the left and right to average}
}
\value{
a vector the same length as vec
}
\description{
Calculating a moving average/max/min statistic over a vector
}
\examples{
movingStat(1:10,mean)
movingStat(1:10,max)
movingStat(1:20,min)
movingStat(rnorm(100),mean,5)
}

