% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlat.R
\name{multiBlatReadsVsRefs}
\alias{multiBlatReadsVsRefs}
\title{Run blatReadsVsRefs in parallel}
\usage{
multiBlatReadsVsRefs(reads, refs, outFile, nCore = 4, tmpDir = tempdir(),
  isGz = grepl(".gz$", outFile), ...)
}
\arguments{
\item{reads}{vector of query reads with names}

\item{refs}{vector of query refs with names}

\item{outFile}{file to write blat to (if ends in .gz then isGz defaults to true and writes to gzipped file)}

\item{nCore}{number of cores to use}

\item{tmpDir}{directory to store work files}

\item{isGz}{if TRUE write to gz file else normal file}

\item{...}{additional arguments for blatReadsVsRefs}
}
\description{
Run blatReadsVsRefs in parallel
}

