% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlat.R
\name{multiRunBlatNoServer}
\alias{multiRunBlatNoServer}
\title{Run blat from parallel blat executables}
\usage{
multiRunBlatNoServer(reads, refs, outFile, nCore = 4,
  tmpDir = tempdir(), isGz = grepl(".gz$", outFile), ...)
}
\arguments{
\item{reads}{named vector of sequences}

\item{refs}{named vector of references}

\item{outFile}{file to write blat to (if ends in .gz then isGz defaults to true and writes to gzipped file)}

\item{nCore}{number of cores to use}

\item{tmpDir}{a directory to write tempfiles to}

\item{isGz}{if TRUE write to gz file else normal file}

\item{...}{additional arguments to runBlatNoServer}
}
\description{
Run blat from parallel blat executables
}
