% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{multiYAxis}
\alias{multiYAxis}
\title{Add multiple text labels to the left axis of a plot}
\usage{
multiYAxis(pos, labs, spacer = 0.3, start = 1, cex = 1.1,
  axisArgs = list(), ...)
}
\arguments{
\item{pos}{position of axis labels in user coordinates}

\item{labs}{a data.frame with each column giving a set of information to display in a column to the left of the plot}

\item{spacer}{amount of space to leave between the widest text in each column. Specified in lines}

\item{start}{where to start the axis labels i.e. the right side of the widest teXt in righmost column. Specified in lines}

\item{cex}{character size for labels}

\item{axisArgs}{a list of additional arguments for axis}

\item{...}{additional arguments to text}
}
\description{
Adds columns of labels to the left axis of a plot e.g. for displaying multiple pieces of information for each row of a heatmap
}
\examples{
par(mar=c(4,10,1,1))
plot(1:10,1:10,ylab='',yaxt='n')
multiYAxis(1:10,data.frame('a'=1:10,'b'=91:100,'c'=1234091:1234100))
}
\seealso{
\code{\link[graphics]{image}}, \code{\link[graphics]{text}}
}
