% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dna.R
\name{noGap2Gap}
\alias{noGap2Gap}
\title{}
\usage{
noGap2Gap(gapSeq, coords)
}
\arguments{
\item{gapSeq}{the reference sequence containing gaps}

\item{coords}{coordinates on the ungapped gapSeq to be converted into equivalent gap coordinates}
}
\value{
equivalent ungapped coordinates
}
\description{

}

