% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{nrmse}
\alias{nrmse}
\title{Calculate the normalized root-mean-square error of two data sets}
\usage{
nrmse(sim_data, obs_data)
}
\arguments{
\item{sim_data}{The simulated data set}

\item{obs_data}{The observed data set (used in the normalization)}
}
\value{
The NRMSE measure.
}
\description{
Use this function to calculate the normalized root-mean-square error
(NRMSE) of two data sets. In this measures, a distinction between the
data sets is needed since the normalization is made using only one of
the data sets. For convention, one of the data sets is called simulated
data set, and the other one is called observed data set.
}
