% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{object.sizes}
\alias{object.sizes}
\title{Convenience function to list objects by size}
\usage{
object.sizes(env = .GlobalEnv)
}
\value{
object sizes in decreasing order
}
\description{
env the environment to list
}

