% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{orderIn}
\alias{orderIn}
\title{Find the  order of one vector in another}
\usage{
orderIn(query, target, strict = FALSE, orderFunc = order, ...)
}
\arguments{
\item{query}{values to be sorted in target order}

\item{target}{order for query to be sorted into}

\item{strict}{if true error if query not in target, if false append unknown queries to end of target}

\item{orderFunc}{function to use for ordering}

\item{...}{arguments for order}
}
\value{
a vector the same length as query specifying the order of query elements in target
}
\description{
Find the  order of one vector in another
}
\examples{
orderIn(c('z','c','a','t'),letters)
}
