% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{parseRegion}
\alias{parseRegion}
\title{Parse a region string into chr, start, end and strand}
\usage{
parseRegion(reg)
}
\arguments{
\item{reg}{vector of region strings in the format "chrX:123545-123324" or "chr12:1234-1236+"}
}
\value{
data.frame with one row per reg string and columns chr, start, end and strand
}
\description{
Parse a region string into chr, start, end and strand
}

