% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{pasteRegion}
\alias{pasteRegion}
\title{Make region from chr start end}
\usage{
pasteRegion(chrs, starts, ends, strands = "")
}
\arguments{
\item{chrs}{vector of chromosomes}

\item{starts}{vector of starts}

\item{ends}{vector of ends}

\item{strands}{optional vector of strands}
}
\value{
vector of region strings
}
\description{
Make region from chr start end
}
\examples{
pasteRegion(c('chr1','chr2'),c(100,1235),c(200,2346))
pasteRegion(c('chr1','chr2'),c(100,1235),c(200,2346),c('*','-'))
}

