% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsd.R
\name{prepare_dsd_4d_module}
\alias{prepare_dsd_4d_module}
\title{Binds the module function with its parameters.}
\usage{
prepare_dsd_4d_module(module, k_idx, species_domains, reactants, products)
}
\arguments{
\item{module}{The module function;}

\item{k_idx}{The k index to name the k for the module;}

\item{species_domains}{The list with the species domains previously
generated}

\item{reactants}{Reactants of the reaction. If a reaction is
something like \code{2A -> B}, a list with
two separated A's \code{('A', 'A')} must be passed.
If a reaction is of formation, \code{NULL} must be passed}

\item{products}{The reaction products. It has the same requirements
of \code{reactants}.}
}
\value{
A function with no arguments. When this function is called, a
string representing the modules with its arguments defined will
be returned.
}
\description{
Use this function to bind a module function to its parameters.
}
