% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{pullRegion}
\alias{pullRegion}
\title{Pull counts from a bam file for target region}
\usage{
pullRegion(reg, files, bam2depthBinary = "./bam2depth",
  fillMissingZeros = TRUE)
}
\arguments{
\item{reg}{region in the format "chrX:123545-123324"}

\item{files}{bam files to pull the counts from}

\item{bam2depthBinary}{bam2depth executable file}

\item{fillMissingZeros}{if TRUE fill in uncovered positions with zeros otherwise may be left out of output by bam2depth}
}
\value{
data.frame with columns chr, pos and counts
}
\description{
Pull counts from a bam file for target region
}
