% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{pwmToSeq}
\alias{pwmToSeq}
\title{Calculate sequence for PWM}
\usage{
pwmToSeq(pwm, cutoff = 0.2)
}
\arguments{
\item{pwm}{a matrix with rows unique bases and columns positions}

\item{cutoff}{consider any base with proportion greater than this cutoff present}
}
\value{
a single character string containing the conensus sequence using ambiguous base codes
}
\description{
Calculate sequence for PWM
}
\examples{
seqs<-c('ACA','ACA','ACA','ACT','ACT','ACC')
seqPwm<-pwm(seqs)
pwmToSeq(seqPwm)
}
\seealso{
\code{\link{pwm}}, \code{\link{ambiguous2regex}}
}
