% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{qualToInts}
\alias{qualToInts}
\title{Convert fastq chars to integer quals}
\usage{
qualToInts(quals, baseQual = 33)
}
\arguments{
\item{quals}{A vector of strings with one character per quality}

\item{baseQual}{integer to subtract from quality characters to convert into quality space}
}
\value{
list with a vector of integer qualities for each string
}
\description{
Convert fastq chars to integer quals
}
\examples{
qualToInts(c('!@#$','ABCI+'))
}
