% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{quickRare}
\alias{quickRare}
\title{Calculate rarefaction using formula}
\usage{
quickRare(sample, step = 10, maxN = sum(sample))
}
\arguments{
\item{sample}{Vector of numbers of individuals per species}

\item{step}{Size of sampling steps for rarefaction}
}
\value{
Matrix with two columns; number of draws and rarefaction value
}
\description{
Calculate rarefaction using formula
}

