% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbow.R
\name{rainbow.lab}
\alias{rainbow.lab}
\title{Rainbow based on circle of lab color space}
\usage{
rainbow.lab(n, start = 1.5, end = -3, alpha = 1, lightScale = 0,
  lightMultiple = 0.7)
}
\arguments{
\item{n}{number of colors desired}

\item{start}{start angle (in proportion of circle) in lab space}

\item{end}{end angle (in proportion of circle) in lab space}

\item{alpha}{transparency (0 for transparent, 1 for opaque)}

\item{lightScale}{exponent to scale lightness by}

\item{lightMultiple}{multiple to scale lightness by}
}
\value{
vector of color strings
}
\description{
Rainbow based on circle of lab color space
}
\examples{
cols<-rainbow.lab(100)
plot(1:100,cex=3,col=cols,lwd=5)
cols<-rainbow.lab(100,lightScale=1)
plot(1:100,col=cols,lwd=5)
}
