% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{rao}
\alias{rao}
\title{Calculate Rao diversity of a vector}
\usage{
rao(x, dist)
}
\arguments{
\item{x}{Vector of counts or proportions of N elements}

\item{dist}{Distance matrix of NxN elements}
}
\value{
Rao diversity
rao(1:3,as.matrix(dist(1:3)))
}
\description{
Calculate Rao diversity of a vector
}

