% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{rarefy}
\alias{rarefy}
\title{Calculate rarefaction using boostrapping}
\usage{
rarefy(counts, samples = unique(round(sum(counts) * seq(0.1, 1, 0.1))),
  reps = 1000, quants = c(0.5, 0.025, 0.975), chaoAdjust = FALSE,
  replace = FALSE)
}
\arguments{
\item{counts}{Counts of species}

\item{samples}{Vector of numbers of draws to calculate rarefaction at}

\item{reps}{How many random samples to take at each step}

\item{quants}{Quantiles to return}

\item{chaoAdjust}{If TRUE calculate chao-predicted species number on each random draw}

\item{replace}{If TRUE sample with replacements. If FALSE sample without replacement.}
}
\value{
Dataframe of calculated quantiles with rownames of the number of samples drawn
}
\description{
Calculate rarefaction using boostrapping
}
\examples{
rarefy(1:20,reps=100)
rarefy(1:20,reps=100,chaoAdjust=TRUE)
}

