% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{rarefyCounts}
\alias{rarefyCounts}
\title{Generate rarefied counts from a vector of counts}
\usage{
rarefyCounts(counts, n)
}
\arguments{
\item{counts}{a vector of counts}

\item{n}{rarefy down to n counts}
}
\value{
a vector of the same length as counts with rarefied counts summing to n
}
\description{
Generate rarefied counts from a vector of counts
}
\examples{
rarefyCounts(1:10,10)
}
\seealso{
\code{link{rarefy}}, \code{link{rareEquation}}
}
