% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{react_4domain}
\alias{react_4domain}
\title{Translate a formal CRN into a set of DNA based reactions according to the
approach described by Soloveichik D et al. \code{[1]}}
\usage{
react_4domain(species, ci, reactions, ki, qmax, cmax, alpha, beta, t)
}
\arguments{
\item{species}{A vector with the species of the reaction. The order of
this vector is important because it will define the
column order of the returned behavior.}

\item{ci}{A vector specifying the initial concentrations of the
\code{species} specified, in order.}

\item{reactions}{A vector with the reactions of the CRN^*.}

\item{ki}{A vector defining the constant rate of each reaction
in \code{reactions}, in order.}

\item{qmax}{Maximum rate constant for the auxiliary reactions.}

\item{cmax}{Maximum initial concentration for the auxiliary species.}

\item{alpha, beta}{Rescaling parameters.}

\item{t}{A vector specifying the time interval. Each value
would be a specific time point.}
}
\value{
A matrix with each line being a specific point in the time
and each column but the first being the concentration of a
species. The first column is the time interval. The
behavior of the auxiliary species are not returned, only
the ones specified in the \code{species} parameter.
}
\description{
This function is used to simulate a chemical circuit based on DNA made
to behavior as expected from a CRN specified by the parameters. In another
words, given the CRN^* passed through the parameters, another CRN_2 is
created based on reactions between strands of DNA. CRN_2 is simulated using
\code{\link{react}()}. The matrix behavior of CRN_2 is returned but only
of the species specified in the \code{species} parameter, the behavior of
the  auxiliary ones are not returned. the parameters of this functions
follows the same pattern of \code{\link{react}()}, only with some additions
required by this approach \code{[1]}.
}
\section{Known limitations}{

\itemize{
\item It only support uni or bimolecular reactions;
\item Because of \code{\link{react}()} known limitation, this function also
doesn't support bidirectional reactions.
}
}

\examples{
library(DNAr)

#
# 4-domain Examples
#

run_ApBeC_4d <- function() {
    behavior <- react_4domain(
        species   = c('A', 'B', 'C'),
        ci        = c(1e3, 1e3, 0),
        reactions = c('A + B -> C'),
        ki        = c(1e-7),
        qmax      = 1e-3,
        cmax      = 1e5,
        alpha     = 1,
        beta      = 1,
        t         = seq(0, 72000, 10)
    )
}

run_AeB_4d <- function() {
    behavior <- react_4domain(
        species   = c('A', 'B'),
        ci        = c(1e4, 0),
        reactions = c('A -> B'),
        ki        = c(5e-5 / 1e5),
        qmax      = 1e-5,
        cmax      = 1e5,
        alpha     = 1,
        beta      = 1,
        t         = seq(0, 72000, 10)
    )
}

run_Lotka_4d <- function() {
    behavior <- react_4domain(
        species   = c('X1', 'X2'),
        ci        = c(20e-9, 10e-9),
        reactions = c('X1 + X2 -> 2X2',
                      'X1 -> 2X1',
                      'X2 -> 0'),
        ki        = c(5e5,
                      1/300,
                      1/300),
        qmax      = 1e6,
        cmax      = 10e-6,
        alpha     = 1,
        beta      = 1,
        t         = seq(0, 12600, 1)
    )
}

run_neuron_4d <- function() {
    behavior <- react_4domain(
        species   = c('X1', 'X2', 'X3', 'X4', 'E1', 'E2', 'I1', 'I2'),
        ci        = c(0, 0, 0, 0, 0.5, 0.5, 10, 5),
        reactions = c('I2 -> 2I2',
                      'I2 -> X3',
                      'I1 -> 2I1',
                      'I1 -> X1',
                      'X1 + E1 -> X2 + E2',
                      'X3 + E2 -> X4 + E1',
                      'X2 -> 0',
                      'X4 -> 0'),
        ki        = c(1, 1, 1, 1, 500, 500, 10, 10),
        qmax      = 10e6,
        cmax      = 10e-5,
        alpha     = 1e3,
        beta      = 5e-7,
        t         = seq(0, 1000, 10)
    )
}

#behavior <- run_ApBeC_4d()
#behavior <- run_AeB_4d()
#behavior <- run_Lotka_4d()
behavior <- run_neuron_4d()

plot_behavior(behavior)
}
\references{
\itemize{
\item \code{[1]} \insertRef{soloveichik2010dna}{DNAr}
}
}
