% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.bed}
\alias{read.bed}
\title{Read a bed file}
\usage{
read.bed(fileName, startAddOne = FALSE)
}
\arguments{
\item{fileName}{.bed file to be read in}

\item{startAddOne}{if TRUE add 1 to starts to adjust for 0-based start, 1-based ends in UCSC formats}
}
\value{
list with a dataframe with columns chr, start and end for each track
}
\description{
Read a bed file
}
