% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.blast}
\alias{read.blast}
\title{Read a tab-delimited blast file}
\usage{
read.blast(fileName, skips = 0, nrows = -1, calcScore = TRUE)
}
\arguments{
\item{fileName}{name of file}

\item{skips}{number of lines to skip (0 for a normal blast file)}

\item{nrows}{number of rows to read in (-1 for all)}

\item{calcScore}{if TRUE, calculate score column}
}
\value{
dataframe of blat data
}
\description{
Read a tab-delimited blast file
}
\examples{
blastString<-c(
  "read1	target1	34.56	100	2	1	101	200	400	500	1e-11	96",
  "read2	target2	100	1000	200	10	102	1100	401	2000	1e-21	123"
)
read.blast(textConnection(blastString))

}
