% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.fa}
\alias{read.fa}
\title{Read a fasta file}
\usage{
read.fa(fileName, assumeSingleLine = FALSE)
}
\arguments{
\item{fileName}{name of file}

\item{assumeSingleLine}{don't process sequence lines. just assume they're one line per sequence}
}
\value{
dataframe with columns name and seq
}
\description{
Read a fasta file
}

