% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.phd}
\alias{read.phd}
\title{Read a sanger phred .phd file}
\usage{
read.phd(fileName, trimQual = -Inf)
}
\arguments{
\item{fileName}{name of file}

\item{trimQual}{trim bases with quality lower than trimQual from start and end}
}
\value{
vector of sequence and space seperated qualities
}
\description{
Read a sanger phred .phd file
}

