% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.sam}
\alias{read.sam}
\title{Read a sam file}
\usage{
read.sam(fileName, nrows = -1, skips = -1, condense = TRUE)
}
\arguments{
\item{fileName}{name of file}

\item{nrows}{number of rows to return. -1 for all rows}

\item{skips}{number of lines to skip (if negative find @ headers automtically)}

\item{condense}{throw out a bunch of columns for smaller file size?}
}
\value{
dataframe with columns qName, flag, tName, pos, cigar, seq
}
\description{
Read a sam file
}

