% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.wiggle}
\alias{read.wiggle}
\title{Read a UCSC wiggle file}
\usage{
read.wiggle(fileName)
}
\arguments{
\item{fileName}{wiggle file to read}
}
\value{
data.frame with columns start, value, end and chr
}
\description{
Read a UCSC wiggle file
}
