% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readAce}
\alias{readAce}
\title{Read an ace file}
\usage{
readAce(aceFile, dropMosaik = TRUE, checkSnps = TRUE, vocal = TRUE)
}
\arguments{
\item{aceFile}{string of file name or file handle}

\item{dropMosaik}{Remove extraneous line from Mosaik labelled either .MosaikAnchor.C1 or MosaikReference}

\item{checkSnps}{Find SNPs from pyroBayes}

\item{vocal}{if TRUE print status messages}
}
\value{
list of reference sequence in [[1]], aligned reads in [[2]] (note reads are not globally aligned still need to use start coordinate to place globally), snps in [[3]]
}
\description{
Read an ace file
}
\section{Warning}{

Only reads 1 contig
}

