% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readBlast}
\alias{readBlast}
\title{Read a tab-delimited blast file}
\usage{
readBlast(fileName, skips = 0, nrows = -1, calcScore = TRUE)
}
\arguments{
\item{fileName}{name of file}

\item{skips}{number of lines to skip (0 for a normal blast file)}

\item{nrows}{number of rows to read in (-1 for all)}

\item{calcScore}{if TRUE, calculate score column}
}
\value{
dataframe of blat data
}
\description{
Read a tab-delimited blast file
}

