% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readBlat}
\alias{readBlat}
\title{Read a blat file}
\usage{
readBlat(fileName, skips = 5, calcScore = TRUE, fixStarts = TRUE,
  nCols = 21, filterFunction = NULL, ...)
}
\arguments{
\item{fileName}{name of file}

\item{skips}{number of lines to skip (5 for a normal blat file)}

\item{calcScore}{calculate score column?}

\item{fixStarts}{convert start to 1-based?}

\item{nCols}{21 (no alignments) or 23 columns (alignments)}

\item{filterFunction}{if not NULL then pass the output data.frame to this function and filter by logical value returned}

\item{...}{additional arguments to read.table}
}
\value{
dataframe of blat data
}
\description{
Read a blat file
}

