% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readLargeBlat}
\alias{readLargeBlat}
\title{Read a large blat file piece by piece}
\usage{
readLargeBlat(
  fileName,
  nHeader = 5,
  nrows = 1e+06,
  isGz = grepl(".gz$", fileName),
  filterFunc = function(x) rep(TRUE, nrow(x)),
  ...
)
}
\arguments{
\item{fileName}{name of file}

\item{nHeader}{number of lines to skip at start of file}

\item{nrows}{number of lines ot read in per shot}

\item{isGz}{if TRUE then file is assumed to be gzipped}

\item{filterFunc}{a function taking a data.frame of blat data and returning a TRUE/FALSE vector specifying rows to keep}

\item{...}{additional arguments to read.blat}
}
\value{
data.frame of blat data
}
\description{
Read a large blat file piece by piece
}
\section{Warning}{

Assumes blat files is sorted by qName
}

