% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{relation_operator}
\alias{relation_operator}
\title{Gets the relation operator of the reaction}
\usage{
relation_operator(reaction)
}
\arguments{
\item{reaction}{A string representing a reaction.}
}
\value{
A string which represents the relation operator of the reaction.
}
\description{
The relation operator is the operator between two parts
of a reaction. Each reaction must have only one operator.
Currently, only the '->' operator is supported.
}
