% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{remove_stoichiometry}
\alias{remove_stoichiometry}
\title{Remove the stoichiometry of a species string.}
\usage{
remove_stoichiometry(species)
}
\value{
The species string without the stoichiometry (number)
specifying the number of molecules
}
\description{
Use this function to get rid of the stoichiometry of a
species string.
}
\examples{
remove_stoichiometry('2A')   # Returns 'A'
remove_stoichiometry('2A2')  # Returns 'A2', The other numbers
                             # are considered part of the species name
}
