% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{replace_module_markers}
\alias{replace_module_markers}
\title{Replaces markers on a template string by other strings}
\usage{
replace_module_markers(template_str, marker_pattern, objs)
}
\arguments{
\item{template_str}{The template string (string with the markers).}

\item{marker_pattern}{The pattern of the markers. It is a regular
expression rule.}

\item{objs}{A vector with the strings that will replace the
markers. The replacement is made in order, that is,
the first markers will be replaced by the first
strings specified in this vector. This vector
must have a number of strings less or equal
the number of markers in the template string.}
}
\value{
The template string with the markers replaced.
}
\description{
This function can be used to replace markers in a string to other
strings according to a pattern.
}
