% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{revComp}
\alias{revComp}
\title{Reverse compliment dna}
\usage{
revComp(dnas)
}
\arguments{
\item{dnas}{vector of sequences}
}
\value{
vector of reverse complimented dna sequences
}
\description{
Reverse compliment dna
}

