% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reverseAmbiguous}
\alias{reverseAmbiguous}
\title{Reverse lookup for ambiguous base codes}
\format{A vector with names giving sortedconcatenated strings of bases for ambiguous base codes with items giving the corresponding ambiguous base code}
\source{
system.file("data-raw", "makeData.R", package = "dnar")
}
\usage{
reverseAmbiguous
}
\description{
Reverse lookup for ambiguous base codes
}
\references{
\url{https://en.wikipedia.org/wiki/Nucleic_acid_notation}
}
\keyword{datasets}

