% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{reverseString}
\alias{reverseString}
\title{Reverse strings}
\usage{
reverseString(strings, brackets = TRUE)
}
\arguments{
\item{strings}{vector of strings to be reversed}

\item{brackets}{if TRUE then reverse brackets and parenthesis, [ goes to ]. keeps brackets in order after reversing}
}
\value{
vector with the reversed strings
}
\description{
Reverse strings
}

