% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rmse}
\alias{rmse}
\title{Calculate the root-mean-square error of two data sets}
\usage{
rmse(data1, data2)
}
\arguments{
\item{data1}{A numerical vector representing one of the data sets.}

\item{data2}{A numerical vector representing the other data set.}
}
\value{
The RMSE value.
}
\description{
This function is used to calculate the root-mean-square error (RMSE)
of two data sets.
}
