% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlat.R
\name{runBlat}
\alias{runBlat}
\title{Start a blat server, run blat on file and kill blat server}
\usage{
runBlat(faFile, gfClientOptions = "", outFile = gsub("\\\\.fn?a$", ".blat",
  faFile), gfClient = "gfClient", ...)
}
\arguments{
\item{faFile}{path to fasta file for alignment}

\item{gfClientOptions}{arguments for gfClient}

\item{outFile}{file to write blat results to}

\item{gfClient}{path to gfClient program}

\item{...}{arguments for startBlat}
}
\description{
Start a blat server, run blat on file and kill blat server
}

