% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlat.R
\name{runBlatNoServer}
\alias{runBlatNoServer}
\title{Run blat from blat executable instead of gfserver/client}
\usage{
runBlatNoServer(reads = NULL, refs, blatArgs = "", outFile = "out.blat",
  blat = "blat", tmpDir = tempdir(), readFile = NULL)
}
\arguments{
\item{reads}{named vector of sequences}

\item{refs}{named vector of references}

\item{blatArgs}{string of arguments for blat}

\item{outFile}{file to write blat to}

\item{blat}{path to blat program}

\item{tmpDir}{a directory to write tempfiles to}

\item{readFile}{file to use instead of reads (will be deleted)}
}
\description{
Run blat from blat executable instead of gfserver/client
}
