% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{samFlag}
\alias{samFlag}
\title{Test sam flags for values}
\usage{
samFlag(flags, test = "paired")
}
\arguments{
\item{flags}{vector of integer flags from sam}

\item{test}{either character vector of flag short names or integers}
}
\value{
logical vector with element for each element in flags with TRUE if flag meets all tests and FALSE otherwise
}
\description{
Test sam flags for values
}

