% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samFlags}
\alias{samFlags}
\title{SAM/BAM flags}
\format{
A data frame with 64 rows one for each codon and columns:
\describe{
  \item{short}{short name for SAM flags}
  \item{desc}{longer description of SAM flags}
  \item{bit}{bit corresponding to the SAM flag}
}
}
\source{
system.file("data-raw", "makeData.R", package = "dnar")
}
\usage{
samFlags
}
\description{
SAM/BAM flags
}
\references{
\url{https://samtools.github.io/hts-specs/SAMv1.pdf}
}
\keyword{datasets}
