% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{samView}
\alias{samView}
\title{Call samtools view on a sam/bam file}
\usage{
samView(fileName, samArgs = "", ..., samtoolsBinary = "samtools",
  vocal = FALSE, samCommand = "view")
}
\arguments{
\item{fileName}{vector of sam/bam files to read}

\item{samArgs}{1 element character vector of args to pass to samtools}

\item{...}{additional arguments for read.sam}

\item{samtoolsBinary}{location of samtools}

\item{vocal}{if TRUE print status messages}

\item{samCommand}{which samtools tool to run}
}
\value{
data.frame with columns as read.sam
}
\description{
Call samtools view on a sam/bam file
}
