% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{save_behavior_csv}
\alias{save_behavior_csv}
\title{Save a reaction behavior}
\usage{
save_behavior_csv(behavior, filename)
}
\arguments{
\item{behavior}{The reaction behavior returned by \code{\link{react}()}.}

\item{filename}{The name of the file that will be saved. You don't need
to specify the extension of the file, the \code{.csv} extension
will be added automatically at the end of the name
specified.}
}
\description{
This function saves a reaction behavior returned by \code{\link{react}()}
in a csv file.
}
