% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{scoreFromPWM}
\alias{scoreFromPWM}
\title{Calculate PWM scores for sequences}
\usage{
scoreFromPWM(seqs, pwm)
}
\arguments{
\item{seqs}{vector of strings, sequences to score}

\item{pwm}{position weight matrix with uniqueBases x seqLength dimensions}
}
\value{
vector with a score for each sequence
}
\description{
Calculate PWM scores for sequences
}
\examples{
seqs<-c('ACA','ACA','ACA','ACT','ACT','ACC')
seqPwm<-pwm(seqs)
scoreFromPWM(seqs,seqPwm)
}

