% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsd.R
\name{select_dsd_4d_module}
\alias{select_dsd_4d_module}
\title{Selects a module based on the reaction stoichiometry}
\usage{
select_dsd_4d_module(n_reac, n_prod)
}
\arguments{
\item{n_reac, n_prod}{Number of reactants and products.}
}
\value{
A list with two parameters:
- \code{mod_func}: The module function obtained from
\code{\link{dsd_4d_modules}()};
- \code{idx}: A vector with two numbers specifying the position
used to get the module function.
}
\description{
This function returns a 4-domain DSD module based on the number
of products and reactants (stoichiometry) of the reaction.
}
