% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/454.R
\name{seq2flow}
\alias{seq2flow}
\title{Convert a DNA sequence into expected 454 flows}
\usage{
seq2flow(seq, flowOrder = c("T", "A", "C", "G"), outputLength = NULL)
}
\arguments{
\item{seq}{DNA sequence}

\item{flowOrder}{order of nucleotide flows}

\item{outputLength}{minimum output length}
}
\value{
vector of flows
}
\description{
Convert a DNA sequence into expected 454 flows
}
\examples{
seq2flow('ACTAAAAAT')
}

