% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{shannon}
\alias{shannon}
\title{Calculate Shannon diversity of a vector}
\usage{
shannon(x, base = exp(1), standardize = TRUE)
}
\arguments{
\item{x}{Vector of counts or proportions}

\item{base}{Base of logarithm}

\item{standardize}{If TRUE divide x by sum x}
}
\value{
Shannon diversity
}
\description{
Calculate Shannon diversity of a vector
}
\examples{
shannon(1:10)
shannon(1:10,base=2)
}
