% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{slantAxis}
\alias{slantAxis}
\title{Label axis with slanted labels}
\usage{
slantAxis(side, at, labels = at, srt = ifelse(side \%in\% c(1, 4), -45,
  45), location = 1.2, adj = ifelse(side == 2, 1, 0),
  axisArgs = list(), ...)
}
\arguments{
\item{side}{an integer specifying which side to draw axis on. 1=bottom, 2=left, 3=top and 4=right.}

\item{at}{locations for ticks and labels}

\item{labels}{character vector of tick labels}

\item{srt}{a numeric specifying string rotation in degrees}

\item{location}{an integer specifying the line to start the text}

\item{adj}{one or two values in [0, 1] which specify the x (and optionally y) adjustment of the labels.  On most devices values outside that interval will also work.}

\item{axisArgs}{a list of additional arguments for axis}

\item{...}{additional arguments to text}
}
\description{
Label axis with slanted labels
}
\examples{
par(mar=c(8,8,8,8))
plot(1:10,xlab='',ylab='',xaxt='n',yaxt='n')
labels<-c('A label','Another label','A longer longer label','A really really\\nlong label')
slantAxis(1,seq(2,8,2),labels)
slantAxis(2,seq(2,8,2),labels)
slantAxis(3,seq(2,8,2),labels)
slantAxis(4,seq(2,8,2),labels,srt=-30,cex=.8,axisArgs=list(col.ticks='red'),lwd=2)
}
