% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stackRegions}
\alias{stackRegions}
\title{Stack regions into smallest number of lines (using greedy algorithm)}
\usage{
stackRegions(starts, ends)
}
\arguments{
\item{starts}{vector of starts of regions (note can add buffer by substracting arbitrary spacer)}

\item{ends}{vector of ends of regions (note can add buffer by adding arbitrary spacer)}
}
\value{
vector of row ids
}
\description{
Stack regions into smallest number of lines (using greedy algorithm)
}
\examples{
stackRegions(1:10,1:10+2)
stackRegions(1:10,1:10+5)
}
