% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlat.R
\name{startBlat}
\alias{startBlat}
\title{Start a blat server}
\usage{
startBlat(nibDir, options = "", gfServer = "gfServer", bit2 = NULL,
  startPort = 37900, nibSuffix = ".nib", wait = 120)
}
\arguments{
\item{nibDir}{directory containing nib files to align against}

\item{options}{options to gfServer}

\item{gfServer}{path to gfServer program}

\item{bit2}{a bit2 file alternative to nib directory}

\item{startPort}{begin looking for open (no other R called gfServer using) ports on startPort and add 1 until finding one}

\item{nibSuffix}{nib files end in nibSuffix}

\item{wait}{number of 5 second intervals to wait for gfServer to finish starting before giving up}
}
\value{
port of blat server
}
\description{
Start a blat server
}

