% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stopError}
\alias{stopError}
\title{Generate an error}
\usage{
stopError(...)
}
\arguments{
\item{...}{Strings to be concatenated into error message}
}
\value{
Generates stopping error before returning
}
\description{
Convenience function for stop(simpleError()
}

