% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{table2vector}
\alias{table2vector}
\title{Convenience function for converting 1-d table to named vector}
\usage{
table2vector(tab)
}
\arguments{
\item{tab}{A one-dimensional table e.g. the output from \code{table}}
}
\value{
a named vector with elements corresponding to the counts in table
}
\description{
Convenience function for converting 1-d table to named vector
}
\examples{
table2vector(1:10)
}

