% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{toggleCase}
\alias{toggleCase}
\title{Swap case of a string}
\usage{
toggleCase(string)
}
\arguments{
\item{string}{String to have case toggled}
}
\value{
String with lowercase and uppercase swapped
}
\description{
Swap case of a string
}

