% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{trim}
\alias{trim}
\title{Trim leading and trailing space characters}
\usage{
trim(x)
}
\arguments{
\item{x}{vector of strings}
}
\value{
vector of trimmed strings
}
\description{
Trim leading and trailing space characters
}

