% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{trimNs}
\alias{trimNs}
\title{Trim ambiguous sequence from ends of reads}
\usage{
trimNs(seqs, start = TRUE, end = TRUE, nonNStretch = NULL,
  nStretch = 10)
}
\arguments{
\item{seqs}{sequences to be trimmed}

\item{start}{trim starts?}

\item{end}{trim ends?}

\item{nonNStretch}{number of nonNs required}

\item{nStretch}{delete until no stretch of Ns greater than this}
}
\value{
Vector sequences with ends trimmed
}
\description{
Trim ambiguous sequence from ends of reads
}

