% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{unifracMatrix}
\alias{unifracMatrix}
\title{Calculate the shared branch length from a matrix of taxonomic labels}
\usage{
unifracMatrix(xx, yy, weighted = TRUE, checkUpstream = TRUE)
}
\arguments{
\item{xx}{a matrix with a row for each read and a column for each taxonomic assignment}

\item{yy}{a matrix with a row for each read and a column for each taxonomic assignment}

\item{weighted}{logical indicating whether to sum the differences in reads for each branch or if FALSE to look only at presence absence}

\item{checkUpstream}{a logical indicating whether to make sure taxa with the same name are counted separetely if their upstream taxa differ e.g. two species Bos taurus and Repipta taurus. If you are sure this is not a problem, then computation can be reduced by skipping this step.}
}
\value{
the proportion of the shared branches
}
\description{
Calculate the shared branch length from a matrix of taxonomic labels
}
\examples{
x<-matrix(c('a','b','a','c'),ncol=2,byrow=TRUE)
y<-matrix(c('a','b','a','d'),ncol=2,byrow=TRUE)
unifracMatrix(x,x)
unifracMatrix(x,y)
}

