% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{update_crn_4domain}
\alias{update_crn_4domain}
\title{Automatically modify a CRN to make it compatible with the DNA simulation}
\usage{
update_crn_4domain(crn, t = NULL, method = NULL)
}
\arguments{
\item{crn}{The CRN to be modified.}

\item{t}{Optional parameter used to redefine the duration time.}

\item{method}{Optional parameter which sets the solver method.}
}
\value{
The modified CRN with the nwe parameters, ready to be used with
the function \code{\link{react_4domain}()}.
}
\description{
This function automatically set the \code{qmax}, \code{cmax}, \code{alpha}, \code{beta}, and
optionally the parameters \code{method} and \code{t} (duration time).
}
