% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wilsonInt}
\alias{wilsonInt}
\title{Calculate the Wilson score interval http://en.wikipedia.org/wiki/Binomial_proportion_confidence_interval}
\usage{
wilsonInt(nTrue, nFalse, alpha = 0.05)
}
\arguments{
\item{nTrue}{number of trues observed}

\item{nFalse}{number of falses observed}

\item{alpha}{error percentile}
}
\value{
two element vector giving the lower and upper bounds of the interval
}
\description{
Calculate the Wilson score interval http://en.wikipedia.org/wiki/Binomial_proportion_confidence_interval
}

