% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{write.fa}
\alias{write.fa}
\title{Writes sequences to a fasta file}
\usage{
write.fa(names, dna, fileName, isGz = grepl(".gz$", fileName))
}
\arguments{
\item{names}{names for the > lines}

\item{dna}{the sequences}

\item{fileName}{file to write to}

\item{isGz}{if TRUE write to a gzipped file}
}
\description{
Writes sequences to a fasta file
}
