% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{write.fastq}
\alias{write.fastq}
\title{Writes a fastq file}
\usage{
write.fastq(names, seqs, quals, fileName, isGz = grepl(".gz$", fileName))
}
\arguments{
\item{names}{sequence names for @/+ line}

\item{seqs}{the sequences strings}

\item{quals}{the qualities strings}

\item{fileName}{file to write to}

\item{isGz}{if TRUE gz compress output}
}
\description{
Writes a fastq file
}
