% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dna.R
\docType{data}
\name{ambiguousBaseCodes}
\alias{ambiguousBaseCodes}
\title{Base codes used to indicate ambiguous bases}
\format{\preformatted{ Named chr [1:11] "AG" "CT" "AC" "GT" "CG" "AT" "ACT" "CGT" ...
 - attr(*, "names")= chr [1:11] "R" "Y" "M" "K" ...
}}
\usage{
ambiguousBaseCodes
}
\description{
Base codes used to indicate ambiguous bases
}
\keyword{datasets}

