% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ambiguousBaseCodes}
\alias{ambiguousBaseCodes}
\title{Ambiguous base codes}
\format{A vector with names corresponding to ambiguous base codes and items giving concatenated strings of corresponding bases}
\source{
system.file("data-raw", "makeData.R", package = "dnar")
}
\usage{
ambiguousBaseCodes
}
\description{
Ambiguous base codes
}
\references{
\url{https://en.wikipedia.org/wiki/Nucleic_acid_notation}
}
\keyword{datasets}
