% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{arrow}
\alias{arrow}
\title{Find coords for arrow plotting}
\usage{
arrow(left, right, y, arrowLength = diff(graphics::par("usr")[1:2]) * 0.05,
  shaft = 0.2, point = 0.4, concat = TRUE)
}
\arguments{
\item{left}{left coordinate of block}

\item{right}{right coordinate of block}

\item{y}{y position for middle of block}

\item{arrowLength}{arrow length in usr coords}

\item{shaft}{half of shaft thickness in usr coords}

\item{point}{half of arrow thickness in usr coords}

\item{concat}{logical indicating whether to concatenate multiple arrows into one data frame separated by NAs (e.g. ready for poly)}
}
\value{
a data.frame with columns x and y specifying coordinates for arrows
}
\description{
Find coords for arrow plotting
}
\examples{
plot(1:10)
polygon(arrow(1:5,3:7*1.5,2:6))
}

