% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{bases2ambiguous}
\alias{bases2ambiguous}
\title{Convert set of single bases to ambiguous code}
\usage{
bases2ambiguous(bases)
}
\arguments{
\item{bases}{vector of character sequences all the same number of characters}
}
\value{
single character string with ambiguous nucleotide codes representing any differences between sequences
}
\description{
Convert set of single bases to ambiguous code
}
\examples{
seqs<-c('ACTAGG','ACGTGG','ACCTGG','ACATGG')
ambig<-bases2ambiguous(seqs)
pattern<-ambiguous2regex(ambig)
grepl(pattern,seqs)
}
\references{
\url{https://en.wikipedia.org/wiki/Nucleic_acid_notation}
}

