% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{blatFindGaps}
\alias{blatFindGaps}
\title{Find gaps in blat coordinates}
\usage{
blatFindGaps(qStarts, tStarts, blockSizes)
}
\arguments{
\item{qStarts}{comma separated starts for query seq}

\item{tStarts}{comma separated starts for target seq}

\item{blockSizes}{comma separated block sizes}
}
\value{
data.frame giving qStartAfter, qGaps, tGapStartAfter and tGaps
}
\description{
Find gaps in blat coordinates
}
\examples{
blatFindGaps('1,100,200','4,200,300','99,50,100')
}
