% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{blatFindGaps}
\alias{blatFindGaps}
\title{Find gaps in blat coordinates}
\usage{
blatFindGaps(qStarts, tStarts, blockSizes)
}
\arguments{
\item{qStarts}{comma seperated starts for query seq}

\item{tStarts}{comma seperated starts for target seq}

\item{blockSizes}{comma seperated block sizes}
}
\value{
data.frame giving qGaps, tGaps, qGapStartAfter and tGapStartAfter
}
\description{
Find gaps in blat coordinates
}

