% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{blockToAlign}
\alias{blockToAlign}
\title{Convert blocks from blat into alignment}
\usage{
blockToAlign(seqs, tSeqs, qStarts, tStarts, sizes)
}
\arguments{
\item{seqs}{vector of sequences}

\item{tSeqs}{vector of target sequences or a single target}

\item{qStarts}{comma separated starts of query matches e.g. from blat or cigarToBlock (1 based)}

\item{tStarts}{comma separated starts of target matches e.g. from blat or cigarToBlock (1 based)}

\item{sizes}{comma separated lengths of matches e.g. from blat or cigarToBlock}
}
\value{
data.frame with sequences aligned in columns qSeq and tSeq
}
\description{
Convert blocks from blat into alignment
}
