% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{check_crn}
\alias{check_crn}
\title{Check if a CRN is respecting the limitations of the function
\code{\link{react}()}}
\usage{
check_crn(species, ci, reactions, ki, t)
}
\description{
This function can be used to check the crn parameters passed to
\code{\link{react}()}. It checks:
\itemize{
\item If all the parameters are correctly defined
and of the correct type;
\item If the length of \code{species} and \code{reactions} are equal to the length
of \code{ci} and \code{ki}, respectively;
\item If there is no duplicate of species names on the \code{species} parameter.
The parameters of this function are the same of \code{\link{react}()}.
}
}
