% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{checkOverlap}
\alias{checkOverlap}
\title{Check overlap between two sets of coordinates}
\usage{
checkOverlap(chrs, starts, ends, tChrs, tStarts, tEnds, tNames, sep = "|")
}
\arguments{
\item{chrs}{Chromosomes of query}

\item{starts}{Start coordinates of query}

\item{ends}{End coordinates of query}

\item{tChrs}{Chromsomes of target}

\item{tStarts}{Start coordinates of target}

\item{tEnds}{End coordinates of target}

\item{tNames}{Target names}

\item{sep}{Matches are pasted together separated by this}
}
\value{
'|' separated vector of tNames within overlap or '' if no overlapping target
}
\description{
Check overlap between two sets of coordinates
}

