% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dna.R
\name{checkOverlap}
\alias{checkOverlap}
\title{Check overlap between two sets of coordinates (ranges package may be better/quicker option)}
\usage{
checkOverlap(starts, ends, tStarts, tEnds, tNames, allCover = FALSE,
  allCoverFuzz = 0, sep = "|")
}
\arguments{
\item{starts}{Start coordinates of query}

\item{ends}{End coordinates of query}

\item{tStarts}{Start coordinates of target}

\item{tEnds}{End coordinates of target}

\item{tNames}{Target names}

\item{allCover}{If TRUE entire start and end of base must fall within targets +- allCoverFuzz}

\item{allCoverFuzz}{Extra overlap to consider on each end of target when using allCover}
}
\value{
'|' seperated vector of tNames within overlap or '' if no overlapping target
}
\description{
Check overlap between two sets of coordinates (ranges package may be better/quicker option)
}

