% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{cigarToBlock}
\alias{cigarToBlock}
\title{Convert cigar and starts to qStarts, tStarts, blockSizes as in blat}
\usage{
cigarToBlock(cigars, starts, startEnds = FALSE, seqs = NULL, tSeq = NULL)
}
\arguments{
\item{cigars}{vector of SAM cigar strings}

\item{starts}{vector of starting positions in target}

\item{startEnds}{if FALSE single line with comma separated starts, if TRUE data.frame with single start, end and id column}

\item{seqs}{vector of query sequences (only if alignments desired)}

\item{tSeq}{single target sequence (only if alignments desired)}
}
\value{
dataframe with qStarts,tStarts,sizes if !startEnds or dataframe with starts, ends and ids if startEnds or dataframe with pairwise query alignments, qAlign, and target alignments, tAlign if provided seqs and tSeq
}
\description{
Convert cigar and starts to qStarts, tStarts, blockSizes as in blat
}

