% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cleanMclapply}
\alias{cleanMclapply}
\title{A function to break an mclapply into parts, save to disk and run independently}
\usage{
cleanMclapply(x, mc.cores, applyFunc, ..., extraCode = "",
  nSplits = mc.cores)
}
\arguments{
\item{x}{vector/list to apply over}

\item{mc.cores}{number of cores to use}

\item{applyFunc}{function to apply}

\item{...}{arguments for mclapply}

\item{extraCode}{character vector of setup code (each command self-contained within a cell or concatenated with ;)}

\item{nSplits}{number of splits to make (if > mc.cores then R will restart more frequently)}
}
\value{
the concatenated outputs from applyFunc
}
\description{
A function to break an mclapply into parts, save to disk and run independently
}

