% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{countNmers}
\alias{countNmers}
\title{Count nMers in a string}
\usage{
countNmers(string, k = 10, n = 10)
}
\arguments{
\item{string}{string to be searched for nmers}

\item{k}{length of nmers}

\item{n}{return the top n nmers}
}
\value{
a sorted table giving the identity and counts for nMers with the highest counts in string
}
\description{
Count nMers in a string
}

