% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cv.glm.par}
\alias{cv.glm.par}
\title{Check predictions of glm by cross validation}
\usage{
cv.glm.par(model, K = nrow(thisData), nCores = 1, subsets = NULL,
  vocal = TRUE)
}
\arguments{
\item{model}{a glm to cross validate}

\item{K}{number of pieces to split data into}

\item{nCores}{number of cores to use}

\item{subsets}{predefined subset of same length as data specifying groupings}

\item{vocal}{echo progress indicator}
}
\value{
a dataframe giving predictions for each data point with columns prediction and subsetId
}
\description{
Check predictions of glm by cross validation
}

