% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dna.R
\name{dna2aa}
\alias{dna2aa}
\title{Convert dna/rna to amino acids}
\usage{
dna2aa(dna, frame = 0, debug = FALSE, ...)
}
\arguments{
\item{dna}{A string of DNA/RNA}

\item{frame}{Starting frame (0=start on first base, 1=on second, 2=on third)}

\item{debug}{If TRUE print debug info}

\item{...}{Additional arguments to codon2aa}
}
\value{
A string of amino acids
}
\description{
Convert dna/rna to amino acids
}

