% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{dna2codons}
\alias{dna2codons}
\title{Convert DNA string into seperate codons}
\usage{
dna2codons(dna, frame = 0)
}
\arguments{
\item{dna}{Vector of DNA strings}

\item{frame}{Starting frame for codons (0=start on first base, 1=on second, 2=on third)}
}
\value{
List with a vector of 3 base codons for each input DNA string
}
\description{
Convert DNA string into seperate codons
}
\examples{
dna2codons('ACATTTGGG')
dna2codons('ACATTTGGG',1)
dna2codons(c('ACATTTGGG','AAATTAGGC'))
dna2codons(c('ACATTTGGG','AAATTAGGC'),c(1,2))
}
