% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{fillZeros}
\alias{fillZeros}
\title{Fill zeros in cover data}
\usage{
fillZeros(cover, posCol = "pos", countCols = colnames(cover)[grep("counts",
  colnames(cover))])
}
\arguments{
\item{cover}{output from pullRegion with missing zero positions}

\item{posCol}{name of position column}

\item{countCols}{vector of column names for columns containing counts
internal}
}
\value{
filled in data.frame
}
\description{
Fill zeros in cover data
}

