% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{generateFakeFasta}
\alias{generateFakeFasta}
\title{Generate fake name and sequence data}
\usage{
generateFakeFasta(nSeq = 10000, nChar = 100:1000, bases = c("A", "C", "T",
  "G", "-", "N"))
}
\arguments{
\item{nSeq}{number of sequences to generate}

\item{nChar}{possible lengths of sequences}

\item{bases}{possible bases}
}
\value{
data.frame with nSeq rows and columns name and seq
}
\description{
Generate fake name and sequence data
}
\examples{
generateFakeFasta(10,10:20)
}

