% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{get_buff_modules}
\alias{get_buff_modules}
\title{Get buffer modules}
\usage{
get_buff_modules(reactions, ki, qmax, cmax)
}
\value{
\code{NULL} if no buffer modules were added. Otherwise
it returns a list with:
\itemize{
\item \code{lambda_1}      = lambda^{-1} value;
\item \code{new_species}   = vector with the new species added;
\item \code{new_cis}       = vector with the initial concentrations;
\item \code{new_reactions} = vector with the new reactions;
\item \code{new_ks}        = contant rate of the new reactions.
}
}
\description{
This function is used to add buffer modules according
to the theory discribed by Soloveichik D et al. \code{[1]}.
The parameters of this function follows the same
semantics of \code{\link{react_4domain}()}.
}
\references{
\itemize{
\item \code{[1]} \insertRef{soloveichik2010dna}{DNAr}
}
}
