% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{get_M}
\alias{get_M}
\title{Returns the matrix M.}
\usage{
get_M(reactions, species)
}
\value{
A list with the named indexes \code{prod} (stoichiometry matrix of the
products), \code{react} (stoichiometry matrix of reactants) and \code{M}
(the M matrix generate from the difference of theses two matrices).
}
\description{
The matrix M is used for the calculations made on \code{\link{react}()}.
This matrix  has #lines = #reactions and #columns = #species and it
represents the stoichiometry of each species at each reaction.
}
