% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_hjelmfelt.R
\name{get_neuron_binding_hje}
\alias{get_neuron_binding_hje}
\title{Get a binding of two neurons}
\usage{
get_neuron_binding_hje(neuron1, neuron2, enzyme_config, ci,
  bind_inhibitory = FALSE)
}
\arguments{
\item{neuron1}{The CRN which defines the neuron 1}

\item{neuron2}{The CRN which defines the neuron 2}

\item{enzyme_config}{A list with the following parameters: \itemize{
\item \code{ci}: the initial concentration of the
enzyme;
\item \code{k}: a vector with the rate constants of
forward and backward reactions, respectively.
}}

\item{ci}{The initial concentration of the C species (the
the binding output which will be used as input of neuron 2).}

\item{bind_inhibitory}{If \code{TRUE}, the inhibitory signal \code{B} will be binded
instead of the activatory one \code{A}.}
}
\value{
The binding CRN.
}
\description{
This function generates a CRN which represents a binding between two neurons
returned by \code{\link{get_neuron_hje}()}. The output of the neuron 1 will be
binded to the input of neuron 2.
}
\details{
The binding is made by creating the reaction \code{E_{ij} + A_{j} -> C_{ij}}.
}
