% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{get_species}
\alias{get_species}
\title{Get the species of a reaction part}
\usage{
get_species(reaction_part)
}
\description{
Given part of a reaction, this function returns
the species of it without the stoichiometry.
}
\examples{
get_species('A + 2B')  # Should return c('A', 'B')
}
