% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dna.R
\name{highlightString}
\alias{highlightString}
\title{Find pattern in strings and flip case}
\usage{
highlightString(pattern, strings)
}
\arguments{
\item{pattern}{Pattern to look for in strings}

\item{strings}{Strings in which to look for pattern}
}
\value{
Strings with case of any occurences of pattern swapped
}
\description{
Find pattern in strings and flip case
}

