% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/454.R
\name{indexFlow}
\alias{indexFlow}
\title{Find flow corresponding to a given base position}
\usage{
indexFlow(flow, coords)
}
\arguments{
\item{flow}{vector of flow values (e.g. produced by seq2flow)}

\item{coords}{1-based coordinates of desired positions}
}
\value{
flow number for each coord. Infinity if coordinate is outside sequence.
}
\description{
Find flow corresponding to a given base position
}
\examples{
indexFlow(seq2flow('CCTTAA'),1:6)
indexFlow(seq2flow('GTGTG'),1:5)
}

