% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{indexMatrix}
\alias{indexMatrix}
\title{Convenience function for selecting multiple elements from a matrix by rows and columns position}
\usage{
indexMatrix(rows, cols, mat, returnIndex = FALSE)
}
\arguments{
\item{rows}{row coordinate of matrix}

\item{cols}{col coorinate of matrix}

\item{mat}{Matrix of interest}

\item{returnIndex}{If TRUE return the one dimensional index for the items otherwise return the selected elements of the matrix}
}
\value{
A vector of indices if returnIndex is TRUE or a vector of the selected matrix elements
}
\description{
Convenience function for selecting multiple elements from a matrix by rows and columns position
}
\examples{
indexMatrix(c(1,2,3),c(1,2,1),matrix(1:9,nrow=3,byrow=TRUE))
}
