% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isError}
\alias{isError}
\title{Convenience function to check for errors in a list of results e.g. from mclapply}
\usage{
isError(x)
}
\arguments{
\item{x}{vector or list to check for errors}
}
\value{
logical vector of length(x) specifying if each element of x was an error
}
\description{
Convenience function to check for errors in a list of results e.g. from mclapply
}
\examples{
isError(c(list(1:3),list(simpleError('Error!'))))
}
\seealso{
\code{\link[parallel]{mclapply}}
}

