% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{noGap2Gap}
\alias{noGap2Gap}
\title{Convert ungapped coordinates to what the coordinates would be with gaps}
\usage{
noGap2Gap(gapSeq, coords)
}
\arguments{
\item{gapSeq}{the reference sequence containing gaps}

\item{coords}{coordinates on the ungapped gapSeq to be converted into equivalent gap coordinates}
}
\value{
equivalent ungapped coordinates
}
\description{
Convert ungapped coordinates to what the coordinates would be with gaps
}

