% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{pRare}
\alias{pRare}
\title{Calculate the probability of observing observedX species from nGroups groups with groupsize members with n observations}
\usage{
pRare(nGroups, groupSize, n, observedX)
}
\arguments{
\item{nGroups}{number of equally sized groups}

\item{groupSize}{size of the equally sized groups}

\item{n}{number of observations pulled from the individuals}

\item{observedX}{number of species observed}
}
\value{
probability of observing observedX species from nGroups groups with groupsize members with n observations
}
\description{
Calculate the probability of observing observedX species from nGroups groups with groupsize members with n observations
}

