% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{rareEquation}
\alias{rareEquation}
\title{Calculate rarefaction using formula}
\usage{
rareEquation(counts, samples = unique(round(sum(counts) * seq(0.1, 1, 0.1))),
  minObs = 1)
}
\arguments{
\item{counts}{Counts of species}

\item{samples}{Vector of numbers of draws to calculate rarefaction at}

\item{minObs}{Minimum number of counts to be counted as present e.g. minObs=2 discards singletons}
}
\value{
Vector with predicted rarefied counts for entries and the sampled number of reads for names
}
\description{
Calculate rarefaction using formula
}
\examples{
rareEquation(1:20)
}
\seealso{
\code{\link{rarefy}}, \code{link{rarefyCounts}}
}
