% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abundance.R
\name{rareEquation}
\alias{rareEquation}
\title{Calculate rarefaction of single sample using equation}
\usage{
rareEquation(speciesCounts, sampleSize)
}
\arguments{
\item{speciesCounts}{Vector of counts for each "species" e.g. c(10,100,5)}

\item{sampleSize}{Single value of number of draws from sample}
}
\value{
Rarefaction value
}
\description{
Calculate rarefaction of single sample using equation
}

