% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{rarefy}
\alias{rarefy}
\title{Calculate rarefaction using boostrapping}
\usage{
rarefy(species, counts = rep(1, length(species)), samples = seq(10,
  sum(counts), 10), reps = 10000, quants = c(0.5, 0.025, 0.975),
  chaoAdjust = FALSE, debug = FALSE, replaceSpecies = FALSE,
  minCount = 0)
}
\arguments{
\item{species}{Ids of species}

\item{counts}{Corresponding counts of species}

\item{samples}{Vector of numbers of draws to calculate rarefaction at}

\item{reps}{How many random samples to take at each step}

\item{quants}{Quantiles to return}

\item{chaoAdjust}{If TRUE calculate chao-predicted species number on each random draw}
}
\value{
Dataframe of calculated quantiles with rownames of the number of samples drawn
}
\description{
Calculate rarefaction using boostrapping
}

