% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{react_4domain}
\alias{react_4domain}
\title{Translate a formal CRN into a set of DNA based reactions according to the
approach described by Soloveichik D et al. \code{[1]}}
\usage{
react_4domain(species, ci, reactions, ki, qmax, cmax, alpha, beta, t)
}
\arguments{
\item{species}{A vector with the species of the reaction. The order of
this vector is important because it will define the
column order of the returned behavior. The species names
\code{L[0-9]*}, \code{H[0-9]*}, \code{W[0-9]*}, \code{O[0-9]*},
\code{T[0-9]*}, \code{G[0-9]*}, \code{LS[0-9]*}, \code{HS[0-9]*},
\code{WS[0-9]*} are not supported. For more information
about this, see the Section of \strong{Known limitations}.}

\item{ci}{A vector specifying the initial concentrations of the
\code{species} specified, in order.}

\item{reactions}{A vector with the reactions of the CRN^*.}

\item{ki}{A vector defining the constant rate of each reaction
in \code{reactions}, in order.}

\item{qmax}{Maximum rate constant for the auxiliary reactions.}

\item{cmax}{Maximum initial concentration for the auxiliary species.}

\item{alpha, beta}{Rescaling parameters.}

\item{t}{A vector specifying the time interval. Each value
would be a specific time point.}
}
\value{
A list with the attributes \code{behavior}, \code{species}, \code{ci}, \code{reactions}
and \code{ki}. These attributes are:
\itemize{
\item \code{behavior}: A matrix with each line being a specific point in the time
and each column but the first being the concentration of a
species. The first column is the time interval. The
behavior of the auxiliary species are also returned.
\item \code{species}  : A vector with all the species used in the reactions.
\item \code{ci}       : The initial concentration of each species.
\item \code{reactions}: All the reactions computed, including the ones generated
according to the 4-domain approach.
\item \code{ki}       : The rate constants of the reactions.
}
}
\description{
This function is used to simulate a chemical circuit based on DNA made
to behavior as expected from a CRN specified by the parameters. In another
words, given the CRN^* passed through the parameters, another \eqn{CRN_2} is
created based on reactions between strands of DNA. CRN_2 is simulated using
\code{\link{react}()}. The matrix behavior of CRN_2 is returned but only
of the species specified in the \code{species} parameter, the behavior of
the  auxiliary ones are not returned. The parameters of these functions
follows the same pattern of \code{\link{react}()}, only with some additions
required by this approach \code{[1]} (here named as 4-domain).
}
\section{Known limitations}{

\itemize{
\item It only support uni or bimolecular reactions;
\item Because of \code{\link{react}()} known limitation, this function also
doesn't support bidirectional reactions;
\item The species names \code{L}, \code{H}, \code{W}, \code{O}, \code{T}, \code{G}, \code{LS},
\code{HS}, \code{WS} with or without numbers after it are not supported because
these are the reserved for the auxiliary ones. Ex.: \code{L2} and \code{LS2}
are not supported but \code{LT} and \code{LT2} are.
}
}

\examples{
library(DNAr)

#
# 4-domain Examples
#

run_ApBeC_4d <- function() {
    result <- react_4domain(
        species   = c('A', 'B', 'C'),
        ci        = c(1e3, 1e3, 0),
        reactions = c('A + B -> C'),
        ki        = c(1e-7),
        qmax      = 1e-3,
        cmax      = 1e5,
        alpha     = 1,
        beta      = 1,
        t         = seq(0, 72000, 10)
    )
    behavior <- result$behavior[,1:(3 + 1)]
}

run_AeB_4d <- function() {
    result <- react_4domain(
        species   = c('A', 'B'),
        ci        = c(1e4, 0),
        reactions = c('A -> B'),
        ki        = c(5e-5 / 1e5),
        qmax      = 1e-5,
        cmax      = 1e5,
        alpha     = 1,
        beta      = 1,
        t         = seq(0, 72000, 10)
    )
    behavior <- result$behavior[,1:(2 + 1)]
}

run_Lotka_4d <- function() {
    result <- react_4domain(
        species   = c('X1', 'X2'),
        ci        = c(20e-9, 10e-9),
        reactions = c('X1 + X2 -> 2X2',
                      'X1 -> 2X1',
                      'X2 -> 0'),
        ki        = c(5e5,
                      1/300,
                      1/300),
        qmax      = 1e6,
        cmax      = 10e-6,
        alpha     = 1,
        beta      = 1,
        t         = seq(0, 12600, 1)
    )
    behavior <- result$behavior[,1:(2 + 1)]
}

#behavior <- run_ApBeC_4d()
#behavior <- run_AeB_4d()
behavior <- run_Lotka_4d()

plot_behavior(behavior)
}
\references{
\itemize{
\item \code{[1]} \insertRef{soloveichik2010dna}{DNAr}
}
}
