% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_reactor.R
\name{reactants_in_reaction}
\alias{reactants_in_reaction}
\title{Check which species are reactants in a given reaction}
\usage{
reactants_in_reaction(species, reaction)
}
\value{
A vector filled with indexes specifying the
species that are in a reaction as a reactant.
}
\description{
It is used to check which of the given species are
reactants in a reaction, returning a vector with the
species indexes in \code{species} that are reactants
in \code{reaction}.
}
\examples{
# Should return c(1, 2)
DNAr:::reactants_in_reaction(c('A', 'B', 'C'), 'A + B -> C')
# Should return c(1)
DNAr:::reactants_in_reaction(c('A', 'B', 'C'), '2A -> B + C')
# Should return c(1, 3)
DNAr:::reactants_in_reaction(c('A', 'B', 'C'), 'A + C -> B')
}
