% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.fa}
\alias{read.fa}
\title{Read a fasta file}
\usage{
read.fa(fileName, assumeSingleLine = FALSE, ...)
}
\arguments{
\item{fileName}{name of file}

\item{assumeSingleLine}{if TRUE don't process sequence lines. just assume they're one line per sequence}

\item{...}{additional arguments to readLines}
}
\value{
data.frame with columns name and seq
}
\description{
Read a fasta file
}
\examples{
file<-tempfile()
x<-generateFasta(100,bases=c('A','C','T','G','\\n','-'))
write.fa(x$name,x$seq,file)
y<-read.fa(file)
}

