% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.fa2}
\alias{read.fa2}
\title{Read a fasta file alternative}
\usage{
read.fa2(fileName, ...)
}
\arguments{
\item{fileName}{name of file}

\item{...}{additional arguments to readLines}
}
\value{
data.frame with columns name and seq
}
\description{
Read a fasta file alternative
}

