% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.fastq}
\alias{read.fastq}
\title{Read fastq file}
\usage{
read.fastq(fileName, convert = FALSE, baseQual = 33)
}
\arguments{
\item{fileName}{name of fastq file}

\item{convert}{if TRUE convert condensed quals to space separated numeric quals}

\item{baseQual}{integer to subtract from quality characters to convert into quality space}
}
\value{
dataframe with name, seq, qual
}
\description{
Read fastq file
}

