% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{read.phd}
\alias{read.phd}
\title{Read a single-read Sanger phred .phd file}
\usage{
read.phd(fileName, trimQual = -Inf)
}
\arguments{
\item{fileName}{name of file}

\item{trimQual}{trim bases with quality lower than trimQual from start and end}
}
\value{
vector of sequence and space seperated qualities
}
\description{
Read a single-read Sanger phred .phd file
}
\examples{
phd<-'Some
extra lines
BEGIN_DNA
c 9 6
t 9 18
c 10 26
c 19 38
END_DNA
more
extra lines'
read.phd(textConnection(phd))
}
