% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{revComp}
\alias{revComp}
\title{Reverse complement dna}
\usage{
revComp(dnas)
}
\arguments{
\item{dnas}{vector of sequences}
}
\value{
vector of reverse complemented dna sequences
}
\description{
Reverse complement dna
}
\examples{
revComp(c('CTAG','ATCCAC'))
revComp(c('CT[AC]G','ATNRY'))
}
