% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4domain_reactor.R
\name{save_dsd_script}
\alias{save_dsd_script}
\title{Export a DSD script for a CRN}
\usage{
save_dsd_script(species, ci, reactions, ki, qmax, cmax, alpha, beta, t,
  filename)
}
\description{
Use this function to export a CRN to a DSD script using the 4-domain
approach of implementing CRNs using DNA. The parameters are the same
of \code{\link{react_4domain}()}, except for the \code{filename}, which is
the name of the file that will be exported and should be specified with
an extension, since there is no standard extension for DSD scripts.
}
