% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dna.R
\name{seqSplit}
\alias{seqSplit}
\title{Convenience function for splitting a bunch of sequences into a matrix}
\usage{
seqSplit(..., fill = NULL)
}
\arguments{
\item{...}{Various sequences to split into a matrix}

\item{fill}{A character to pad ends of sequences. Sequences are left uneven lengths if NULL}
}
\value{
A matrix of single characters each row corresponding to a read
}
\description{
Convenience function for splitting a bunch of sequences into a matrix
}

